@@ -5,17 +5,18 @@ Based on https://github.com/topper-123/optioneer, but simplified (don't deal
 with nested options, deprecated options, ..), just the attribute-style dict
 like holding the options and giving a nice repr.
 """
-from collections import namedtuple
-import textwrap
 
+import textwrap
+import warnings
+from collections import namedtuple
 
 Option = namedtuple("Option", "key default_value doc validator callback")
 
 
-class Options(object):
+class Options:
     """Provide attribute-style access to configuration dict."""
 
-    def __init__(self, options):
+    def __init__(self, options: dict[str, Option]) -> None:
         super().__setattr__("_options", options)
         # populate with default values
         config = {}
@@ -24,7 +25,7 @@ class Options(object):
 
         super().__setattr__("_config", config)
 
-    def __setattr__(self, key, value):
+    def __setattr__(self, key, value) -> None:
         # you can't set new keys
         if key in self._config:
             option = self._options[key]
@@ -43,15 +44,15 @@ class Options(object):
         except KeyError:
             raise AttributeError("No such option")
 
-    def __dir__(self):
+    def __dir__(self) -> list[str]:
         return list(self._config.keys())
 
-    def __repr__(self):
+    def __repr__(self) -> str:
         cls = self.__class__.__name__
         description = ""
         for key, option in self._options.items():
-            descr = "{key}: {cur!r} [default: {default!r}]\n".format(
-                key=key, cur=self._config[key], default=option.default_value
+            descr = (
+                f"{key}: {self._config[key]!r} [default: {option.default_value!r}]\n"
             )
             description += descr
 
@@ -63,10 +64,10 @@ class Options(object):
             description += doc_text + "\n"
         space = "\n  "
         description = description.replace("\n", space)
-        return "{}({}{})".format(cls, space, description)
+        return f"{cls}({space}{description})"
 
 
-def _validate_display_precision(value):
+def _validate_display_precision(value: int) -> None:
     if value is not None:
         if not isinstance(value, int) or not (0 <= value <= 16):
             raise ValueError("Invalid value, needs to be an integer [0-16]")
@@ -86,35 +87,35 @@ display_precision = Option(
 )
 
 
-def _validate_bool(value):
-    if not isinstance(value, bool):
-        raise TypeError("Expected bool value, got {0}".format(type(value)))
+def _warn_use_pygeos_deprecated(_value) -> None:
+    warnings.warn(
+        "pygeos support was removed in 1.0. "
+        "geopandas.use_pygeos is a no-op and will be removed in geopandas 1.1.",
+        stacklevel=3,
+    )
 
 
-def _default_use_pygeos():
-    import geopandas._compat as compat
-
-    return compat.USE_PYGEOS
-
-
-def _callback_use_pygeos(key, value):
-    assert key == "use_pygeos"
-    import geopandas._compat as compat
-
-    compat.set_use_pygeos(value)
+def _validate_io_engine(value: str) -> None:
+    if value is not None:
+        if value not in ("pyogrio", "fiona"):
+            raise ValueError(f"Expected 'pyogrio' or 'fiona', got '{value}'")
 
 
-use_pygeos = Option(
-    key="use_pygeos",
-    default_value=_default_use_pygeos(),
+io_engine = Option(
+    key="io_engine",
+    default_value=None,
     doc=(
-        "Whether to use PyGEOS to speed up spatial operations. The default is True "
-        "if PyGEOS is installed, and follows the USE_PYGEOS environment variable "
-        "if set."
+        "The default engine for ``read_file`` and ``to_file``. "
+        "Options are 'pyogrio' and 'fiona'."
     ),
-    validator=_validate_bool,
-    callback=_callback_use_pygeos,
+    validator=_validate_io_engine,
+    callback=None,
 )
 
 
-options = Options({"display_precision": display_precision, "use_pygeos": use_pygeos})
+options = Options(
+    {
+        "display_precision": display_precision,
+        "io_engine": io_engine,
+    }
+)
