@@ -1,8 +1,12 @@
+from __future__ import annotations
+
+import inspect
 import numbers
 import operator
+import typing
 import warnings
-import inspect
 from functools import lru_cache
+from typing import Any, ClassVar, Literal
 
 import numpy as np
 import pandas as pd
@@ -15,41 +19,70 @@ from pandas.api.extensions import (
 import shapely
 import shapely.affinity
 import shapely.geometry
-from shapely.geometry.base import BaseGeometry
 import shapely.ops
 import shapely.wkt
-from pyproj import CRS, Transformer
+from shapely.geometry.base import BaseGeometry
 
-try:
-    import pygeos
-except ImportError:
-    geos = None
+from ._compat import (
+    GEOS_GE_312,
+    HAS_PYPROJ,
+    requires_pyproj,
+)
+from .sindex import SpatialIndex
 
-from . import _compat as compat
-from . import _vectorized as vectorized
-from .sindex import _get_sindex_class
+if typing.TYPE_CHECKING:
+    import numpy.typing as npt
+
+    from .base import GeoPandasBase
+
+    if HAS_PYPROJ:
+        from pyproj import CRS
+    else:
+        CRS = Any
+
+if HAS_PYPROJ:
+    from pyproj import Transformer
+
+    TransformerFromCRS = lru_cache(Transformer.from_crs)
 
 
-TransformerFromCRS = lru_cache(Transformer.from_crs)
+_names = {
+    "MISSING": None,
+    "NAG": None,
+    "POINT": "Point",
+    "LINESTRING": "LineString",
+    "LINEARRING": "LinearRing",
+    "POLYGON": "Polygon",
+    "MULTIPOINT": "MultiPoint",
+    "MULTILINESTRING": "MultiLineString",
+    "MULTIPOLYGON": "MultiPolygon",
+    "GEOMETRYCOLLECTION": "GeometryCollection",
+}
+
+POLYGON_GEOM_TYPES = {"Polygon", "MultiPolygon"}
+LINE_GEOM_TYPES = {"LineString", "MultiLineString", "LinearRing"}
+POINT_GEOM_TYPES = {"Point", "MultiPoint"}
+
+type_mapping = {p.value: _names[p.name] for p in shapely.GeometryType}
+geometry_type_ids = list(type_mapping.keys())
+geometry_type_values = np.array(list(type_mapping.values()), dtype=object)
 
 
 class GeometryDtype(ExtensionDtype):
     type = BaseGeometry
     name = "geometry"
-    na_value = np.nan
+    na_value = None
 
     @classmethod
     def construct_from_string(cls, string):
         if not isinstance(string, str):
             raise TypeError(
-                "'construct_from_string' expects a string, got {}".format(type(string))
+                f"'construct_from_string' expects a string, got {type(string)}"
             )
         elif string == cls.name:
             return cls()
         else:
-            raise TypeError(
-                "Cannot construct a '{}' from '{}'".format(cls.__name__, string)
-            )
+            raise TypeError(f"Cannot construct a '{cls.__name__}' from '{string}'")
 
     @classmethod
     def construct_array_type(cls):
@@ -59,7 +92,9 @@ class GeometryDtype(ExtensionDtype):
 register_extension_dtype(GeometryDtype)
 
 
-def _check_crs(left, right, allow_none=False):
+def _check_crs(
+    left: GeoPandasBase, right: GeoPandasBase, allow_none: bool = False
+) -> bool:
     """
     Check if the projection of both arrays is the same.
 
@@ -73,10 +108,10 @@ def _check_crs(left, right, allow_none=False):
     return True
 
 
-def _crs_mismatch_warn(left, right, stacklevel=3):
-    """
-    Raise a CRS mismatch warning with the information on the assigned CRS.
-    """
+def _crs_mismatch_warn(
+    left: GeoPandasBase, right: GeoPandasBase, stacklevel: int = 3
+) -> None:
+    """Raise a CRS mismatch warning with the information on the assigned CRS."""
     if left.crs:
         left_srs = left.crs.to_string()
         left_srs = left_srs if len(left_srs) <= 50 else " ".join([left_srs[:50], "..."])
@@ -96,11 +131,28 @@ def _crs_mismatch_warn(left, right, stacklevel=3):
         "and the CRS of right geometries.\n"
         "Use `to_crs()` to reproject one of "
         "the input geometries to match the CRS of the other.\n\n"
-        "Left CRS: {0}\n"
-        "Right CRS: {1}\n".format(left_srs, right_srs),
+        f"Left CRS: {left_srs}\n"
+        f"Right CRS: {right_srs}\n",
         UserWarning,
         stacklevel=stacklevel,
     )
+
+
+def isna(value: None | float | pd.NA) -> bool:
+    """
+    Check if scalar value is NA-like (None, np.nan or pd.NA).
+
+    Custom version that only works for scalars (returning True or False),
+    as `pd.isna` also works for array-like input returning a boolean array.
+    """
+    if value is None:
+        return True
+    elif isinstance(value, float) and np.isnan(value):
+        return True
+    elif value is pd.NA:
+        return True
+    else:
+        return False
 
 
 # -----------------------------------------------------------------------------
@@ -108,38 +160,11 @@ def _crs_mismatch_warn(left, right, stacklevel=3):
 # -----------------------------------------------------------------------------
 
 
-def _geom_to_shapely(geom):
-    """
-    Convert internal representation (PyGEOS or Shapely) to external Shapely object.
-    """
-    if compat.USE_SHAPELY_20:
-        return geom
-    elif not compat.USE_PYGEOS:
-        return geom
-    else:
-        return vectorized._pygeos_to_shapely(geom)
+def _is_scalar_geometry(geom) -> bool:
+    return isinstance(geom, BaseGeometry)
 
 
-def _shapely_to_geom(geom):
-    """
-    Convert external Shapely object to internal representation (PyGEOS or Shapely).
-    """
-    if compat.USE_SHAPELY_20:
-        return geom
-    elif not compat.USE_PYGEOS:
-        return geom
-    else:
-        return vectorized._shapely_to_pygeos(geom)
-
-
-def _is_scalar_geometry(geom):
-    if compat.USE_PYGEOS:
-        return isinstance(geom, (pygeos.Geometry, BaseGeometry))
-    else:
-        return isinstance(geom, BaseGeometry)
-
-
-def from_shapely(data, crs=None):
+def from_shapely(data, crs: CRS | None = None) -> GeometryArray:
     """
     Convert a list or array of shapely objects to a GeometryArray.
 
@@ -155,19 +180,44 @@ def from_shapely(data, crs=None):
         such as an authority string (eg "EPSG:4326") or a WKT string.
 
     """
-    return GeometryArray(vectorized.from_shapely(data), crs=crs)
+    if not isinstance(data, np.ndarray):
+        arr = np.empty(len(data), dtype=object)
+        arr[:] = data
+    elif len(data) == 0 and data.dtype == "float64":
+        arr = data.astype(object)
+    else:
+        arr = data
+
+    if not shapely.is_valid_input(arr).all():
+        out = []
+
+        for geom in data:
+            if isinstance(geom, BaseGeometry):
+                out.append(geom)
+            elif hasattr(geom, "__geo_interface__"):
+                geom = shapely.geometry.shape(geom)
+                out.append(geom)
+            elif isna(geom):
+                out.append(None)
+            else:
+                raise TypeError(f"Input must be valid geometry objects: {geom}")
+        arr = np.array(out, dtype=object)
+
+    return GeometryArray(arr, crs=crs)
 
 
-def to_shapely(geoms):
-    """
-    Convert GeometryArray to numpy object array of shapely objects.
-    """
+def to_shapely(geoms: GeometryArray) -> np.ndarray:
+    """Convert GeometryArray to numpy object array of shapely objects."""
     if not isinstance(geoms, GeometryArray):
         raise ValueError("'geoms' must be a GeometryArray")
-    return vectorized.to_shapely(geoms._data)
+    return geoms._data
 
 
-def from_wkb(data, crs=None):
+def from_wkb(
+    data,
+    crs: Any | None = None,
+    on_invalid: Literal["raise", "warn", "ignore"] = "raise",
+) -> GeometryArray:
     """
     Convert a list or array of WKB objects to a GeometryArray.
 
@@ -179,21 +229,33 @@ def from_wkb(data, crs=None):
         Coordinate Reference System of the geometry objects. Can be anything accepted by
         :meth:`pyproj.CRS.from_user_input() <pyproj.crs.CRS.from_user_input>`,
         such as an authority string (eg "EPSG:4326") or a WKT string.
+    on_invalid: {"raise", "warn", "ignore"}, default "raise"
+        - raise: an exception will be raised if a WKB input geometry is invalid.
+        - warn: a warning will be raised and invalid WKB geometries will be returned as
+          None.
+        - ignore: invalid WKB geometries will be returned as None without a warning.
+        - fix: an effort is made to fix invalid input geometries (e.g. close
+          unclosed rings). If this is not possible, they are returned as ``None``
+          without a warning. Requires GEOS >= 3.11 and shapely >= 2.1.
 
     """
-    return GeometryArray(vectorized.from_wkb(data), crs=crs)
+    if isinstance(data, ExtensionArray):
+        data = data.to_numpy(na_value=None)
+    return GeometryArray(shapely.from_wkb(data, on_invalid=on_invalid), crs=crs)
 
 
-def to_wkb(geoms, hex=False, **kwargs):
-    """
-    Convert GeometryArray to a numpy object array of WKB objects.
-    """
+def to_wkb(geoms: GeometryArray, hex: bool = False, **kwargs):
+    """Convert GeometryArray to a numpy object array of WKB objects."""
     if not isinstance(geoms, GeometryArray):
         raise ValueError("'geoms' must be a GeometryArray")
-    return vectorized.to_wkb(geoms._data, hex=hex, **kwargs)
+    return shapely.to_wkb(geoms, hex=hex, **kwargs)
 
 
-def from_wkt(data, crs=None):
+def from_wkt(
+    data,
+    crs: Any | None = None,
+    on_invalid: Literal["raise", "warn", "ignore"] = "raise",
+) -> GeometryArray:
     """
     Convert a list or array of WKT objects to a GeometryArray.
 
@@ -205,21 +267,31 @@ def from_wkt(data, crs=None):
         Coordinate Reference System of the geometry objects. Can be anything accepted by
         :meth:`pyproj.CRS.from_user_input() <pyproj.crs.CRS.from_user_input>`,
         such as an authority string (eg "EPSG:4326") or a WKT string.
+    on_invalid : {"raise", "warn", "ignore"}, default "raise"
+        - raise: an exception will be raised if a WKT input geometry is invalid.
+        - warn: a warning will be raised and invalid WKT geometries will be
+          returned as ``None``.
+        - ignore: invalid WKT geometries will be returned as ``None`` without a warning.
+        - fix: an effort is made to fix invalid input geometries (e.g. close
+          unclosed rings). If this is not possible, they are returned as ``None``
+          without a warning. Requires GEOS >= 3.11 and shapely >= 2.1.
 
     """
-    return GeometryArray(vectorized.from_wkt(data), crs=crs)
+    if isinstance(data, ExtensionArray):
+        data = data.to_numpy(na_value=None)
+    return GeometryArray(shapely.from_wkt(data, on_invalid=on_invalid), crs=crs)
 
 
-def to_wkt(geoms, **kwargs):
-    """
-    Convert GeometryArray to a numpy object array of WKT objects.
-    """
+def to_wkt(geoms: GeometryArray, **kwargs):
+    """Convert GeometryArray to a numpy object array of WKT objects."""
     if not isinstance(geoms, GeometryArray):
         raise ValueError("'geoms' must be a GeometryArray")
-    return vectorized.to_wkt(geoms._data, **kwargs)
+    return shapely.to_wkt(geoms, **kwargs)
 
 
-def points_from_xy(x, y, z=None, crs=None):
+def points_from_xy(
+    x: npt.ArrayLike, y: npt.ArrayLike, z: npt.ArrayLike = None, crs: Any | None = None
+) -> GeometryArray:
     """
     Generate GeometryArray of shapely Point geometries from x, y(, z) coordinates.
 
@@ -262,18 +334,23 @@ def points_from_xy(x, y, z=None, crs=None):
     -------
     output : GeometryArray
     """
-    return GeometryArray(vectorized.points_from_xy(x, y, z), crs=crs)
+    x = np.asarray(x, dtype="float64")
+    y = np.asarray(y, dtype="float64")
+    if z is not None:
+        z = np.asarray(z, dtype="float64")
+
+    return GeometryArray(shapely.points(x, y, z), crs=crs)
 
 
 class GeometryArray(ExtensionArray):
-    """
-    Class wrapping a numpy array of Shapely objects and
-    holding the array-based implementations.
+    """Class wrapping a numpy array of Shapely objects.
+
+    It also holds the array-based implementations.
     """
 
     _dtype = GeometryDtype()
 
-    def __init__(self, data, crs=None):
+    def __init__(self, data, crs: Any | None = None):
         if isinstance(data, self.__class__):
             if not crs:
                 crs = data.crs
@@ -294,29 +371,14 @@ class GeometryArray(ExtensionArray):
         self._sindex = None
 
     @property
-    def data(self):
-        warnings.warn(
-            "Accessing the underlying geometries through the `.data` attribute is "
-            "deprecated and will be removed in GeoPandas 1.0. You can use "
-            "`np.asarray(..)` or the `to_numpy()` method instead.\n"
-            "Note that if you are using PyGEOS and using this attribute to get an "
-            "array of PyGEOS geometries, those other methods will always return an "
-            "array of Shapely geometries. Accessing the underlying PyGEOS geometries "
-            "directly is deprecated, and you should migrate to use Shapely >= 2.0 "
-            "instead.",
-            DeprecationWarning,
-            stacklevel=2,
-        )
-        return self._data
-
-    @property
-    def sindex(self):
+    def sindex(self) -> SpatialIndex:
+        """Spatial index for the geometries in this array."""
         if self._sindex is None:
-            self._sindex = _get_sindex_class()(self._data)
+            self._sindex = SpatialIndex(self._data)
         return self._sindex
 
     @property
-    def has_sindex(self):
+    def has_sindex(self) -> bool:
         """Check the existence of the spatial index without generating it.
 
         Use the `.sindex` attribute on a GeoDataFrame or GeoSeries
@@ -328,7 +390,7 @@ class GeometryArray(ExtensionArray):
         initialized until the first use.
 
         See Also
-        ---------
+        --------
         GeoDataFrame.has_sindex
 
         Returns
@@ -340,13 +402,10 @@ class GeometryArray(ExtensionArray):
         return self._sindex is not None
 
     @property
-    def crs(self):
-        """
-        The Coordinate Reference System (CRS) represented as a ``pyproj.CRS``
-        object.
+    def crs(self) -> CRS:
+        """The Coordinate Reference System (CRS) represented as a ``pyproj.CRS`` object.
 
-        Returns None if the CRS is not set, and to set the value it
-        :getter: Returns a ``pyproj.CRS`` or None. When setting, the value
+        Returns a ``pyproj.CRS`` or None. When setting, the value
         Coordinate Reference System of the geometry objects. Can be anything accepted by
         :meth:`pyproj.CRS.from_user_input() <pyproj.crs.CRS.from_user_input>`,
         such as an authority string (eg "EPSG:4326") or a WKT string.
@@ -354,33 +413,58 @@ class GeometryArray(ExtensionArray):
         return self._crs
 
     @crs.setter
-    def crs(self, value):
-        """Sets the value of the crs"""
-        self._crs = None if not value else CRS.from_user_input(value)
+    def crs(self, value: Any) -> None:
+        """Set the value of the crs."""
+        if HAS_PYPROJ:
+            from pyproj import CRS
 
-    def check_geographic_crs(self, stacklevel):
-        """Check CRS and warn if the planar operation is done in a geographic CRS"""
+            self._crs = None if not value else CRS.from_user_input(value)
+        else:
+            if value is not None:
+                warnings.warn(
+                    "Cannot set the CRS, falling back to None. The CRS support requires"
+                    " the 'pyproj' package, but it is not installed or does not import"
+                    " correctly. The functions depending on CRS will raise an error or"
+                    " may produce unexpected results.",
+                    UserWarning,
+                    stacklevel=2,
+                )
+            self._crs = None
+
+    def check_geographic_crs(self, stacklevel: int) -> None:
+        """Check CRS and warn if the planar operation is done in a geographic CRS."""
         if self.crs and self.crs.is_geographic:
             warnings.warn(
-                "Geometry is in a geographic CRS. Results from '{}' are likely "
-                "incorrect. Use 'GeoSeries.to_crs()' to re-project geometries to a "
-                "projected CRS before this operation.\n".format(
-                    inspect.stack()[1].function
-                ),
+                "Geometry is in a geographic CRS. Results from "
+                f"'{inspect.stack()[1].function}' are likely incorrect. "
+                "Use 'GeoSeries.to_crs()' to re-project geometries to a "
+                "projected CRS before this operation.\n",
                 UserWarning,
                 stacklevel=stacklevel,
             )
 
     @property
-    def dtype(self):
+    def dtype(self) -> GeometryDtype:
         return self._dtype
 
-    def __len__(self):
+    def __len__(self) -> int:
         return self.shape[0]
 
-    def __getitem__(self, idx):
+    def __getitem__(self, idx) -> GeometryArray:
         if isinstance(idx, numbers.Integral):
-            return _geom_to_shapely(self._data[idx])
+            return self._data[idx]
+        elif (
+            isinstance(idx, slice)
+            and idx.start is None
+            and idx.stop is None
+            and idx.step is None
+        ):
+            # special case of a full slice -> preserve the sindex
+            # (to ensure view() preserves it as well)
+            result = GeometryArray(self._data[idx], crs=self.crs)
+            result._sindex = self._sindex
+            return result
+
         # array-like, slice
         # validate and convert IntegerArray/BooleanArray
         # to numpy array, pass-through non-array-like indexers
@@ -395,14 +479,14 @@ class GeometryArray(ExtensionArray):
             value = value.values
         if isinstance(value, pd.DataFrame):
             value = value.values.flatten()
-        if isinstance(value, (list, np.ndarray)):
+        if isinstance(value, list | np.ndarray):
             value = from_shapely(value)
         if isinstance(value, GeometryArray):
             if isinstance(key, numbers.Integral):
                 raise ValueError("cannot set a single element with an array")
             self._data[key] = value._data
-        elif isinstance(value, BaseGeometry) or vectorized.isna(value):
-            if vectorized.isna(value):
+        elif isinstance(value, BaseGeometry) or isna(value):
+            if isna(value):
                 # internally only use None as missing value indicator
                 # but accept others
                 value = None
@@ -410,16 +494,15 @@ class GeometryArray(ExtensionArray):
                 value = from_shapely([value])._data[0]
             else:
                 raise TypeError("should be valid geometry")
-            if isinstance(key, (slice, list, np.ndarray)):
+            if isinstance(key, slice | list | np.ndarray):
                 value_array = np.empty(1, dtype=object)
-                with compat.ignore_shapely2_warnings():
-                    value_array[:] = [value]
+                value_array[:] = [value]
                 self._data[key] = value_array
             else:
                 self._data[key] = value
         else:
             raise TypeError(
-                "Value should be either a BaseGeometry or None, got %s" % str(value)
+                f"Value should be either a BaseGeometry or None, got {value!s}"
             )
 
         # invalidate spatial index
@@ -434,17 +517,12 @@ class GeometryArray(ExtensionArray):
         #         )
 
     def __getstate__(self):
-        if compat.USE_SHAPELY_20:
-            return (shapely.to_wkb(self._data), self._crs)
-        elif compat.USE_PYGEOS:
-            return (pygeos.to_wkb(self._data), self._crs)
-        else:
-            return self.__dict__
+        return (shapely.to_wkb(self._data), self._crs)
 
     def __setstate__(self, state):
         if not isinstance(state, dict):
             # pickle file saved with pygeos
-            geoms = vectorized.from_wkb(state[0])
+            geoms = shapely.from_wkb(state[0])
             self._crs = state[1]
             self._sindex = None  # pygeos.STRtree could not be pickled yet
             self._data = geoms
@@ -452,8 +530,6 @@ class GeometryArray(ExtensionArray):
         else:
             if "data" in state:
                 state["_data"] = state.pop("data")
-            if compat.USE_PYGEOS:
-                state["_data"] = vectorized.from_shapely(state["_data"])
             if "_crs" not in state:
                 state["_crs"] = None
             self.__dict__.update(state)
@@ -464,88 +540,250 @@ class GeometryArray(ExtensionArray):
 
     @property
     def is_valid(self):
-        return vectorized.is_valid(self._data)
+        return shapely.is_valid(self._data)
+
+    def is_valid_reason(self):
+        return shapely.is_valid_reason(self._data)
+
+    def is_valid_coverage(self, gap_width: float = 0.0):
+        if not GEOS_GE_312:
+            raise ImportError("Method 'is_valid_coverage' requires GEOS>=3.12.")
+        return bool(shapely.coverage_is_valid(self._data, gap_width=gap_width))
+
+    def invalid_coverage_edges(self, gap_width: float = 0.0):
+        if not GEOS_GE_312:
+            raise ImportError(
+                "Method 'invalid_coverage_edges' requires and GEOS>=3.12."
+            )
+        return shapely.coverage_invalid_edges(self._data, gap_width=gap_width)
 
     @property
     def is_empty(self):
-        return vectorized.is_empty(self._data)
+        return shapely.is_empty(self._data)
 
     @property
     def is_simple(self):
-        return vectorized.is_simple(self._data)
+        return shapely.is_simple(self._data)
 
     @property
     def is_ring(self):
-        return vectorized.is_ring(self._data)
+        return shapely.is_ring(self._data)
 
     @property
     def is_closed(self):
-        return vectorized.is_closed(self._data)
+        return shapely.is_closed(self._data)
+
+    @property
+    def is_ccw(self):
+        return shapely.is_ccw(self._data)
 
     @property
     def has_z(self):
-        return vectorized.has_z(self._data)
+        return shapely.has_z(self._data)
+
+    @property
+    def has_m(self):
+        return shapely.has_m(self._data)
 
     @property
     def geom_type(self):
-        return vectorized.geom_type(self._data)
+        res = shapely.get_type_id(self._data)
+        return geometry_type_values[np.searchsorted(geometry_type_ids, res)]
 
     @property
     def area(self):
+        """Return the area of the geometries in this array.
+
+        Raises a UserWarning if the CRS is geographic, as the area
+        calculation is not accurate in that case.
+
+        Note that the area is calculated in the units of the CRS.
+
+        Returns
+        -------
+        np.ndarray of float
+            Area of the geometries.
+        """
         self.check_geographic_crs(stacklevel=5)
-        return vectorized.area(self._data)
+        return shapely.area(self._data)
 
     @property
     def length(self):
         self.check_geographic_crs(stacklevel=5)
-        return vectorized.length(self._data)
+        return shapely.length(self._data)
+
+    def count_coordinates(self):
+        return shapely.get_num_coordinates(self._data)
+
+    def count_geometries(self):
+        return shapely.get_num_geometries(self._data)
+
+    def count_interior_rings(self):
+        return shapely.get_num_interior_rings(self._data)
+
+    def get_precision(self):
+        return shapely.get_precision(self._data)
+
+    def get_geometry(self, index):
+        return shapely.get_geometry(self._data, index=index)
 
     #
     # Unary operations that return new geometries
     #
 
     @property
-    def boundary(self):
-        return GeometryArray(vectorized.boundary(self._data), crs=self.crs)
+    def boundary(self) -> GeometryArray:
+        return GeometryArray(shapely.boundary(self._data), crs=self.crs)
 
     @property
-    def centroid(self):
+    def centroid(self) -> GeometryArray:
         self.check_geographic_crs(stacklevel=5)
-        return GeometryArray(vectorized.centroid(self._data), crs=self.crs)
+        return GeometryArray(shapely.centroid(self._data), crs=self.crs)
 
-    @property
-    def convex_hull(self):
-        return GeometryArray(vectorized.convex_hull(self._data), crs=self.crs)
+    def concave_hull(self, ratio, allow_holes):
+        return shapely.concave_hull(self._data, ratio=ratio, allow_holes=allow_holes)
 
-    @property
-    def envelope(self):
-        return GeometryArray(vectorized.envelope(self._data), crs=self.crs)
-
-    @property
-    def exterior(self):
-        return GeometryArray(vectorized.exterior(self._data), crs=self.crs)
-
-    @property
-    def interiors(self):
-        # no GeometryArray as result
-        return vectorized.interiors(self._data)
-
-    def representative_point(self):
-        return GeometryArray(vectorized.representative_point(self._data), crs=self.crs)
-
-    def minimum_bounding_circle(self):
+    def constrained_delaunay_triangles(self) -> GeometryArray:
         return GeometryArray(
-            vectorized.minimum_bounding_circle(self._data), crs=self.crs
+            shapely.constrained_delaunay_triangles(self._data), crs=self.crs
+        )
+
+    @property
+    def convex_hull(self) -> GeometryArray:
+        """Return the convex hull of the geometries in this array."""
+        return GeometryArray(shapely.convex_hull(self._data), crs=self.crs)
+
+    @property
+    def envelope(self) -> GeometryArray:
+        """Return the envelope of the geometries in this array."""
+        return GeometryArray(shapely.envelope(self._data), crs=self.crs)
+
+    def minimum_rotated_rectangle(self):
+        """Return the minimum rotated rectangle of the geometries in this array."""
+        return GeometryArray(shapely.oriented_envelope(self._data), crs=self.crs)
+
+    @property
+    def exterior(self) -> GeometryArray:
+        return GeometryArray(shapely.get_exterior_ring(self._data), crs=self.crs)
+
+    def extract_unique_points(self) -> GeometryArray:
+        return GeometryArray(shapely.extract_unique_points(self._data), crs=self.crs)
+
+    def offset_curve(self, distance, quad_segs=8, join_style="round", mitre_limit=5.0):
+        return GeometryArray(
+            shapely.offset_curve(
+                self._data,
+                distance,
+                quad_segs=quad_segs,
+                join_style=join_style,
+                mitre_limit=mitre_limit,
+            ),
+            crs=self.crs,
+        )
+
+    @property
+    def interiors(self) -> np.ndarray:
+        # no GeometryArray as result
+        has_non_poly = False
+        inner_rings = []
+        for geom in self._data:
+            interior_ring_seq = getattr(geom, "interiors", None)
+            # polygon case
+            if interior_ring_seq is not None:
+                inner_rings.append(list(interior_ring_seq))
+            # non-polygon case
+            else:
+                has_non_poly = True
+                inner_rings.append(None)
+        if has_non_poly:
+            warnings.warn(
+                "Only Polygon objects have interior rings. For other "
+                "geometry types, None is returned.",
+                stacklevel=2,
+            )
+        # need to allocate empty first in case of all empty lists in inner_rings
+        data = np.empty(len(inner_rings), dtype=object)
+        data[:] = inner_rings
+        return data
+
+    def remove_repeated_points(self, tolerance=0.0) -> GeometryArray:
+        return GeometryArray(
+            shapely.remove_repeated_points(self._data, tolerance=tolerance),
+            crs=self.crs,
+        )
+
+    def representative_point(self) -> GeometryArray:
+        return GeometryArray(shapely.point_on_surface(self._data), crs=self.crs)
+
+    def minimum_bounding_circle(self) -> GeometryArray:
+        return GeometryArray(shapely.minimum_bounding_circle(self._data), crs=self.crs)
+
+    def maximum_inscribed_circle(self, tolerance) -> GeometryArray:
+        return GeometryArray(
+            shapely.maximum_inscribed_circle(self._data, tolerance=tolerance),
+            crs=self.crs,
         )
 
     def minimum_bounding_radius(self):
-        return vectorized.minimum_bounding_radius(self._data)
+        return shapely.minimum_bounding_radius(self._data)
 
-    def normalize(self):
-        return GeometryArray(vectorized.normalize(self._data), crs=self.crs)
+    def minimum_clearance(self):
+        return shapely.minimum_clearance(self._data)
 
-    def make_valid(self):
-        return GeometryArray(vectorized.make_valid(self._data), crs=self.crs)
+    def minimum_clearance_line(self) -> GeometryArray:
+        return GeometryArray(shapely.minimum_clearance_line(self._data), crs=self.crs)
+
+    def normalize(self) -> GeometryArray:
+        return GeometryArray(shapely.normalize(self._data), crs=self.crs)
+
+    def orient_polygons(self, exterior_cw: bool = False) -> GeometryArray:
+        return GeometryArray(
+            shapely.orient_polygons(self._data, exterior_cw=exterior_cw), crs=self.crs
+        )
+
+    def make_valid(
+        self,
+        method: Literal["linework", "structure"] = "linework",
+        keep_collapsed: bool = True,
+    ) -> GeometryArray:
+        return GeometryArray(
+            shapely.make_valid(
+                self._data, method=method, keep_collapsed=keep_collapsed
+            ),
+            crs=self.crs,
+        )
+
+    def reverse(self) -> GeometryArray:
+        return GeometryArray(shapely.reverse(self._data), crs=self.crs)
+
+    def segmentize(self, max_segment_length) -> GeometryArray:
+        return GeometryArray(
+            shapely.segmentize(self._data, max_segment_length),
+            crs=self.crs,
+        )
+
+    def force_2d(self) -> GeometryArray:
+        return GeometryArray(shapely.force_2d(self._data), crs=self.crs)
+
+    def force_3d(self, z=0) -> GeometryArray:
+        return GeometryArray(shapely.force_3d(self._data, z=z), crs=self.crs)
+
+    def transform(self, transformation, include_z: bool = False) -> GeometryArray:
+        return GeometryArray(
+            shapely.transform(self._data, transformation, include_z=include_z),
+            crs=self.crs,
+        )
+
+    def line_merge(self, directed: bool = False) -> GeometryArray:
+        return GeometryArray(
+            shapely.line_merge(self._data, directed=directed), crs=self.crs
+        )
+
+    def set_precision(self, grid_size: float, mode="valid_output"):
+        return GeometryArray(
+            shapely.set_precision(self._data, grid_size=grid_size, mode=mode),
+            crs=self.crs,
+        )
 
     #
     # Binary predicates
@@ -555,15 +793,16 @@ class GeometryArray(ExtensionArray):
     def _binary_method(op, left, right, **kwargs):
         if isinstance(right, GeometryArray):
             if len(left) != len(right):
-                msg = "Lengths of inputs do not match. Left: {0}, Right: {1}".format(
-                    len(left), len(right)
+                msg = (
+                    "Lengths of inputs do not match. "
+                    f"Left: {len(left)}, Right: {len(right)}"
                 )
                 raise ValueError(msg)
             if not _check_crs(left, right):
                 _crs_mismatch_warn(left, right, stacklevel=7)
             right = right._data
 
-        return getattr(vectorized, op)(left._data, right, **kwargs)
+        return getattr(shapely, op)(left._data, right, **kwargs)
 
     def covers(self, other):
         return self._binary_method("covers", self, other)
@@ -573,6 +812,9 @@ class GeometryArray(ExtensionArray):
 
     def contains(self, other):
         return self._binary_method("contains", self, other)
+
+    def contains_properly(self, other):
+        return self._binary_method("contains_properly", self, other)
 
     def crosses(self, other):
         return self._binary_method("crosses", self, other)
@@ -595,39 +837,64 @@ class GeometryArray(ExtensionArray):
     def within(self, other):
         return self._binary_method("within", self, other)
 
+    def dwithin(self, other, distance):
+        self.check_geographic_crs(stacklevel=6)
+        return self._binary_method("dwithin", self, other, distance=distance)
+
     def geom_equals_exact(self, other, tolerance):
         return self._binary_method("equals_exact", self, other, tolerance=tolerance)
 
-    def geom_almost_equals(self, other, decimal):
-        return self.geom_equals_exact(other, 0.5 * 10 ** (-decimal))
-        # return _binary_predicate("almost_equals", self, other, decimal=decimal)
+    def geom_equals_identical(self, other):
+        return self._binary_method("equals_identical", self, other)
 
     #
     # Binary operations that return new geometries
     #
 
-    def clip_by_rect(self, xmin, ymin, xmax, ymax):
+    def clip_by_rect(self, xmin, ymin, xmax, ymax) -> GeometryArray:
         return GeometryArray(
-            vectorized.clip_by_rect(self._data, xmin, ymin, xmax, ymax), crs=self.crs
+            shapely.clip_by_rect(self._data, xmin, ymin, xmax, ymax), crs=self.crs
         )
 
-    def difference(self, other):
+    def difference(self, other, grid_size=None) -> GeometryArray:
         return GeometryArray(
-            self._binary_method("difference", self, other), crs=self.crs
+            self._binary_method("difference", self, other, grid_size=grid_size),
+            crs=self.crs,
         )
 
-    def intersection(self, other):
+    def intersection(self, other, grid_size=None) -> GeometryArray:
         return GeometryArray(
-            self._binary_method("intersection", self, other), crs=self.crs
+            self._binary_method("intersection", self, other, grid_size=grid_size),
+            crs=self.crs,
         )
 
-    def symmetric_difference(self, other):
+    def symmetric_difference(self, other, grid_size=None) -> GeometryArray:
         return GeometryArray(
-            self._binary_method("symmetric_difference", self, other), crs=self.crs
+            self._binary_method(
+                "symmetric_difference", self, other, grid_size=grid_size
+            ),
+            crs=self.crs,
         )
 
-    def union(self, other):
-        return GeometryArray(self._binary_method("union", self, other), crs=self.crs)
+    def union(self, other, grid_size=None) -> GeometryArray:
+        return GeometryArray(
+            self._binary_method("union", self, other, grid_size=grid_size), crs=self.crs
+        )
+
+    def shortest_line(self, other) -> GeometryArray:
+        return GeometryArray(
+            self._binary_method("shortest_line", self, other), crs=self.crs
+        )
+
+    def snap(self, other, tolerance) -> GeometryArray:
+        return GeometryArray(
+            self._binary_method("snap", self, other, tolerance=tolerance), crs=self.crs
+        )
+
+    def shared_paths(self, other) -> GeometryArray:
+        return GeometryArray(
+            self._binary_method("shared_paths", self, other), crs=self.crs
+        )
 
     #
     # Other operations
@@ -637,91 +904,198 @@ class GeometryArray(ExtensionArray):
         self.check_geographic_crs(stacklevel=6)
         return self._binary_method("distance", self, other)
 
-    def buffer(self, distance, resolution=16, **kwargs):
-        if not (isinstance(distance, (int, float)) and distance == 0):
+    def hausdorff_distance(self, other, **kwargs):
+        self.check_geographic_crs(stacklevel=6)
+        return self._binary_method("hausdorff_distance", self, other, **kwargs)
+
+    def frechet_distance(self, other, **kwargs):
+        self.check_geographic_crs(stacklevel=6)
+        return self._binary_method("frechet_distance", self, other, **kwargs)
+
+    def buffer(self, distance, quad_segs: int | None = None, **kwargs):
+        if not (isinstance(distance, int | float) and distance == 0):
             self.check_geographic_crs(stacklevel=5)
+        if "resolution" in kwargs:
+            if quad_segs is not None:
+                msg = (
+                    "`buffer` received both `quad_segs` and `resolution` but these are "
+                    "aliases for the same parameter. Use `quad_segs` only instead."
+                )
+                raise ValueError(msg)
+
+            msg = (
+                "The `resolution` argument to `buffer` is deprecated, `quad_segs` "
+                "should be used instead to align with shapely."
+            )
+            warnings.warn(
+                msg,
+                category=DeprecationWarning,
+                stacklevel=4,
+            )
+            quad_segs = kwargs.pop("resolution")
+        if quad_segs is None:
+            quad_segs = 16  # note shapely default is 8, 16 is historical choice
+
         return GeometryArray(
-            vectorized.buffer(self._data, distance, resolution=resolution, **kwargs),
+            shapely.buffer(self._data, distance, quad_segs=quad_segs, **kwargs),
             crs=self.crs,
         )
 
-    def interpolate(self, distance, normalized=False):
+    def interpolate(self, distance, normalized: bool = False) -> GeometryArray:
         self.check_geographic_crs(stacklevel=5)
         return GeometryArray(
-            vectorized.interpolate(self._data, distance, normalized=normalized),
+            shapely.line_interpolate_point(self._data, distance, normalized=normalized),
             crs=self.crs,
         )
 
-    def simplify(self, tolerance, preserve_topology=True):
+    def simplify(self, tolerance, preserve_topology: bool = True) -> GeometryArray:
         return GeometryArray(
-            vectorized.simplify(
+            shapely.simplify(
                 self._data, tolerance, preserve_topology=preserve_topology
             ),
             crs=self.crs,
         )
 
+    def simplify_coverage(
+        self, tolerance, simplify_boundary: bool = True
+    ) -> GeometryArray:
+        if not GEOS_GE_312:
+            raise ImportError(
+                "'simplify_coverage' requires shapely>=2.1 and GEOS>=3.12."
+            )
+        return GeometryArray(
+            shapely.coverage_simplify(
+                self._data, tolerance, simplify_boundary=simplify_boundary
+            ),
+            crs=self.crs,
+        )
+
     def project(self, other, normalized=False):
-        if isinstance(other, BaseGeometry):
-            other = _shapely_to_geom(other)
-        elif isinstance(other, GeometryArray):
+        if isinstance(other, GeometryArray):
             other = other._data
-        return vectorized.project(self._data, other, normalized=normalized)
+        return shapely.line_locate_point(self._data, other, normalized=normalized)
 
     def relate(self, other):
         if isinstance(other, GeometryArray):
             other = other._data
-        return vectorized.relate(self._data, other)
+        return shapely.relate(self._data, other)
+
+    def relate_pattern(self, other, pattern):
+        if isinstance(other, GeometryArray):
+            other = other._data
+        return shapely.relate_pattern(self._data, other, pattern)
 
     #
     # Reduction operations that return a Shapely geometry
     #
 
     def unary_union(self):
-        return vectorized.unary_union(self._data)
+        warnings.warn(
+            "The 'unary_union' attribute is deprecated, "
+            "use the 'union_all' method instead.",
+            DeprecationWarning,
+            stacklevel=2,
+        )
+        return self.union_all()
+
+    def union_all(self, method="unary", grid_size=None):
+        if method != "unary" and grid_size is not None:
+            raise ValueError(f"grid_size is not supported for method '{method}'.")
+        if method == "coverage":
+            return shapely.coverage_union_all(self._data)
+        elif method == "unary":
+            return shapely.union_all(self._data, grid_size=grid_size)
+        elif method == "disjoint_subset":
+            if not GEOS_GE_312:
+                raise ImportError(
+                    "Method 'disjoin_subset' requires shapely>=2.1 and GEOS>=3.12."
+                )
+            return shapely.disjoint_subset_union_all(self._data)
+        else:
+            raise ValueError(
+                f"Method '{method}' not recognized. Use 'coverage', 'unary' or "
+                "'disjoint_subset'."
+            )
+
+    def intersection_all(self):
+        return shapely.intersection_all(self._data)
 
     #
     # Affinity operations
     #
 
-    def affine_transform(self, matrix):
+    @staticmethod
+    def _affinity_method(op, left, *args, **kwargs):
+        # not all shapely.affinity methods can handle empty geometries:
+        # affine_transform itself works (as well as translate), but rotate, scale
+        # and skew fail (they try to unpack the bounds).
+        # Here: consistently returning empty geom for input empty geom
+        out = []
+        for geom in left:
+            if geom is None or geom.is_empty:
+                res = geom
+            else:
+                res = getattr(shapely.affinity, op)(geom, *args, **kwargs)
+            out.append(res)
+        data = np.empty(len(left), dtype=object)
+        data[:] = out
+        return data
+
+    def affine_transform(self, matrix) -> GeometryArray:
         return GeometryArray(
-            vectorized._affinity_method("affine_transform", self._data, matrix),
+            self._affinity_method("affine_transform", self._data, matrix),
             crs=self.crs,
         )
 
-    def translate(self, xoff=0.0, yoff=0.0, zoff=0.0):
+    def translate(
+        self, xoff: float = 0.0, yoff: float = 0.0, zoff: float = 0.0
+    ) -> GeometryArray:
         return GeometryArray(
-            vectorized._affinity_method("translate", self._data, xoff, yoff, zoff),
+            self._affinity_method("translate", self._data, xoff, yoff, zoff),
             crs=self.crs,
         )
 
-    def rotate(self, angle, origin="center", use_radians=False):
+    def rotate(
+        self, angle, origin="center", use_radians: bool = False
+    ) -> GeometryArray:
         return GeometryArray(
-            vectorized._affinity_method(
+            self._affinity_method(
                 "rotate", self._data, angle, origin=origin, use_radians=use_radians
             ),
             crs=self.crs,
         )
 
-    def scale(self, xfact=1.0, yfact=1.0, zfact=1.0, origin="center"):
+    def scale(
+        self,
+        xfact: float = 1.0,
+        yfact: float = 1.0,
+        zfact: float = 1.0,
+        origin="center",
+    ) -> GeometryArray:
         return GeometryArray(
-            vectorized._affinity_method(
+            self._affinity_method(
                 "scale", self._data, xfact, yfact, zfact, origin=origin
             ),
             crs=self.crs,
         )
 
-    def skew(self, xs=0.0, ys=0.0, origin="center", use_radians=False):
+    def skew(
+        self,
+        xs: float = 0.0,
+        ys: float = 0.0,
+        origin="center",
+        use_radians: bool = False,
+    ) -> GeometryArray:
         return GeometryArray(
-            vectorized._affinity_method(
+            self._affinity_method(
                 "skew", self._data, xs, ys, origin=origin, use_radians=use_radians
             ),
             crs=self.crs,
         )
 
-    def to_crs(self, crs=None, epsg=None):
-        """Returns a ``GeometryArray`` with all geometries transformed to a new
-        coordinate reference system.
+    @requires_pyproj
+    def to_crs(self, crs: Any | None = None, epsg: int | None = None) -> GeometryArray:
+        """Transform all geometries to a different coordinate reference system.
 
         Transform all geometries in a GeometryArray to a different coordinate
         reference system.  The ``crs`` attribute on the current GeometryArray must
@@ -788,6 +1162,8 @@ class GeometryArray(ExtensionArray):
         - Prime Meridian: Greenwich
 
         """
+        from pyproj import CRS
+
         if self.crs is None:
             raise ValueError(
                 "Cannot transform naive geometries.  "
@@ -806,11 +1182,12 @@ class GeometryArray(ExtensionArray):
 
         transformer = TransformerFromCRS(self.crs, crs, always_xy=True)
 
-        new_data = vectorized.transform(self._data, transformer.transform)
+        new_data = transform(self._data, transformer.transform)
         return GeometryArray(new_data, crs=crs)
 
-    def estimate_utm_crs(self, datum_name="WGS 84"):
-        """Returns the estimated UTM CRS based on the bounds of the dataset.
+    @requires_pyproj
+    def estimate_utm_crs(self, datum_name: str = "WGS 84") -> CRS:
+        """Return the estimated UTM CRS based on the bounds of the dataset.
 
         .. versionadded:: 0.9
 
@@ -827,31 +1204,29 @@ class GeometryArray(ExtensionArray):
 
         Examples
         --------
-        >>> world = geopandas.read_file(
-        ...     geopandas.datasets.get_path("naturalearth_lowres")
+        >>> import geodatasets
+        >>> df = geopandas.read_file(
+        ...     geodatasets.get_path("geoda.chicago_commpop")
         ... )
-        >>> germany = world.loc[world.name == "Germany"]
-        >>> germany.geometry.values.estimate_utm_crs()  # doctest: +SKIP
-        <Projected CRS: EPSG:32632>
-        Name: WGS 84 / UTM zone 32N
+        >>> df.geometry.values.estimate_utm_crs()  # doctest: +SKIP
+        <Derived Projected CRS: EPSG:32616>
+        Name: WGS 84 / UTM zone 16N
         Axis Info [cartesian]:
         - E[east]: Easting (metre)
         - N[north]: Northing (metre)
         Area of Use:
-        - name: World - N hemisphere - 6°E to 12°E - by country
-        - bounds: (6.0, 0.0, 12.0, 84.0)
+        - name: Between 90°W and 84°W, northern hemisphere between equator and 84°N,...
+        - bounds: (-90.0, 0.0, -84.0, 84.0)
         Coordinate Operation:
-        - name: UTM zone 32N
+        - name: UTM zone 16N
         - method: Transverse Mercator
-        Datum: World Geodetic System 1984
+        Datum: World Geodetic System 1984 ensemble
         - Ellipsoid: WGS 84
         - Prime Meridian: Greenwich
         """
-        try:
-            from pyproj.aoi import AreaOfInterest
-            from pyproj.database import query_utm_crs_info
-        except ImportError:
-            raise RuntimeError("pyproj 3+ required for estimate_utm_crs.")
+        from pyproj import CRS
+        from pyproj.aoi import AreaOfInterest
+        from pyproj.database import query_utm_crs_info
 
         if not self.crs:
             raise RuntimeError("crs must be set to estimate UTM CRS.")
@@ -863,28 +1238,21 @@ class GeometryArray(ExtensionArray):
         # ensure using geographic coordinates
         else:
             transformer = TransformerFromCRS(self.crs, "EPSG:4326", always_xy=True)
-            if compat.PYPROJ_GE_31:
-                minx, miny, maxx, maxy = transformer.transform_bounds(
-                    minx, miny, maxx, maxy
-                )
-                y_center = np.mean([miny, maxy])
-                # crossed the antimeridian
-                if minx > maxx:
-                    # shift maxx from [-180,180] to [0,360]
-                    # so both numbers are positive for center calculation
-                    # Example: -175 to 185
-                    maxx += 360
-                    x_center = np.mean([minx, maxx])
-                    # shift back to [-180,180]
-                    x_center = ((x_center + 180) % 360) - 180
-                else:
-                    x_center = np.mean([minx, maxx])
+            minx, miny, maxx, maxy = transformer.transform_bounds(
+                minx, miny, maxx, maxy
+            )
+            y_center = np.mean([miny, maxy])
+            # crossed the antimeridian
+            if minx > maxx:
+                # shift maxx from [-180,180] to [0,360]
+                # so both numbers are positive for center calculation
+                # Example: -175 to 185
+                maxx += 360
+                x_center = np.mean([minx, maxx])
+                # shift back to [-180,180]
+                x_center = ((x_center + 180) % 360) - 180
             else:
-                lon, lat = transformer.transform(
-                    (minx, maxx, minx, maxx), (miny, miny, maxy, maxy)
-                )
-                x_center = np.mean(lon)
-                y_center = np.mean(lat)
+                x_center = np.mean([minx, maxx])
 
         utm_crs_list = query_utm_crs_info(
             datum_name=datum_name,
@@ -906,55 +1274,71 @@ class GeometryArray(ExtensionArray):
 
     @property
     def x(self):
-        """Return the x location of point geometries in a GeoSeries"""
+        """Return the x location of point geometries in a GeoSeries."""
         if (self.geom_type[~self.isna()] == "Point").all():
             empty = self.is_empty
             if empty.any():
                 nonempty = ~empty
                 coords = np.full_like(nonempty, dtype=float, fill_value=np.nan)
-                coords[nonempty] = vectorized.get_x(self._data[nonempty])
+                coords[nonempty] = shapely.get_x(self._data[nonempty])
                 return coords
             else:
-                return vectorized.get_x(self._data)
+                return shapely.get_x(self._data)
         else:
             message = "x attribute access only provided for Point geometries"
             raise ValueError(message)
 
     @property
     def y(self):
-        """Return the y location of point geometries in a GeoSeries"""
+        """Return the y location of point geometries in a GeoSeries."""
         if (self.geom_type[~self.isna()] == "Point").all():
             empty = self.is_empty
             if empty.any():
                 nonempty = ~empty
                 coords = np.full_like(nonempty, dtype=float, fill_value=np.nan)
-                coords[nonempty] = vectorized.get_y(self._data[nonempty])
+                coords[nonempty] = shapely.get_y(self._data[nonempty])
                 return coords
             else:
-                return vectorized.get_y(self._data)
+                return shapely.get_y(self._data)
         else:
             message = "y attribute access only provided for Point geometries"
             raise ValueError(message)
 
     @property
     def z(self):
-        """Return the z location of point geometries in a GeoSeries"""
+        """Return the z location of point geometries in a GeoSeries."""
         if (self.geom_type[~self.isna()] == "Point").all():
             empty = self.is_empty
             if empty.any():
                 nonempty = ~empty
                 coords = np.full_like(nonempty, dtype=float, fill_value=np.nan)
-                coords[nonempty] = vectorized.get_z(self._data[nonempty])
+                coords[nonempty] = shapely.get_z(self._data[nonempty])
                 return coords
             else:
-                return vectorized.get_z(self._data)
+                return shapely.get_z(self._data)
         else:
             message = "z attribute access only provided for Point geometries"
             raise ValueError(message)
 
     @property
+    def m(self):
+        """Return the m coordinate of point geometries in a GeoSeries."""
+        if (self.geom_type[~self.isna()] == "Point").all():
+            empty = self.is_empty
+            if empty.any():
+                nonempty = ~empty
+                coords = np.full_like(nonempty, dtype=float, fill_value=np.nan)
+                coords[nonempty] = shapely.get_m(self._data[nonempty])
+                return coords
+            else:
+                return shapely.get_m(self._data)
+        else:
+            message = "m attribute access only provided for Point geometries"
+            raise ValueError(message)
+
+    @property
     def bounds(self):
-        return vectorized.bounds(self._data)
+        return shapely.bounds(self._data)
 
     @property
     def total_bounds(self):
@@ -963,14 +1347,19 @@ class GeometryArray(ExtensionArray):
             # TODO with numpy >= 1.15, the 'initial' argument can be used
             return np.array([np.nan, np.nan, np.nan, np.nan])
         b = self.bounds
-        return np.array(
-            (
-                np.nanmin(b[:, 0]),  # minx
-                np.nanmin(b[:, 1]),  # miny
-                np.nanmax(b[:, 2]),  # maxx
-                np.nanmax(b[:, 3]),  # maxy
+        with warnings.catch_warnings():
+            # if all rows are empty geometry / none, nan is expected
+            warnings.filterwarnings(
+                "ignore", r"All-NaN slice encountered", RuntimeWarning
             )
-        )
+            return np.array(
+                (
+                    np.nanmin(b[:, 0]),  # minx
+                    np.nanmin(b[:, 1]),  # miny
+                    np.nanmax(b[:, 2]),  # maxx
+                    np.nanmax(b[:, 3]),  # maxy
+                )
+            )
 
     # -------------------------------------------------------------------------
     # general array like compat
@@ -981,59 +1370,41 @@ class GeometryArray(ExtensionArray):
         return self._data.size
 
     @property
-    def shape(self):
+    def shape(self) -> tuple:
         return (self.size,)
 
     @property
-    def ndim(self):
+    def ndim(self) -> int:
         return len(self.shape)
 
-    def copy(self, *args, **kwargs):
+    def copy(self, *args, **kwargs) -> GeometryArray:
         # still taking args/kwargs for compat with pandas 0.24
         return GeometryArray(self._data.copy(), crs=self._crs)
 
-    def take(self, indices, allow_fill=False, fill_value=None):
+    def take(self, indices, allow_fill: bool = False, fill_value=None) -> GeometryArray:
         from pandas.api.extensions import take
 
         if allow_fill:
             if fill_value is None or pd.isna(fill_value):
                 fill_value = None
-            elif isinstance(fill_value, BaseGeometry):
-                fill_value = _shapely_to_geom(fill_value)
             elif not _is_scalar_geometry(fill_value):
                 raise TypeError("provide geometry or None as fill value")
 
         result = take(self._data, indices, allow_fill=allow_fill, fill_value=fill_value)
         if allow_fill and fill_value is None:
-            result[pd.isna(result)] = None
+            result[~shapely.is_valid_input(result)] = None
         return GeometryArray(result, crs=self.crs)
 
-    def _fill(self, idx, value):
-        """
-        Fill index locations with ``value``.
+    # compat for pandas < 3.0
+    def _pad_or_backfill(
+        self, method, limit=None, limit_area=None, copy=True, **kwargs
+    ):
+        kwargs["limit_area"] = limit_area
+        return super()._pad_or_backfill(method=method, limit=limit, copy=copy, **kwargs)
 
-        ``value`` should be a BaseGeometry or a GeometryArray.
-        """
-        if vectorized.isna(value):
-            value = [None]
-        elif _is_scalar_geometry(value):
-            value = [value]
-        elif isinstance(value, GeometryArray):
-            value = value[idx]
-        else:
-            raise TypeError(
-                "'value' parameter must be None, a scalar geometry, or a GeoSeries, "
-                f"but you passed a {type(value).__name__!r}"
-            )
-
-        value_arr = np.empty(len(value), dtype=object)
-        with compat.ignore_shapely2_warnings():
-            value_arr[:] = _shapely_to_geom(value)
-
-        self._data[idx] = value_arr
-        return self
-
-    def fillna(self, value=None, method=None, limit=None):
+    def fillna(
+        self, value=None, method=None, limit: int | None = None, copy: bool = True
+    ) -> GeometryArray:
         """
         Fill NA values with geometry (or geometries) or using the specified method.
 
@@ -1041,7 +1412,7 @@ class GeometryArray(ExtensionArray):
         ----------
         value : shapely geometry object or GeometryArray
             If a geometry value is passed it is used to fill all missing values.
-            Alternatively, an GeometryArray 'value' can be given. It's expected
+            Alternatively, a GeometryArray 'value' can be given. It's expected
             that the GeometryArray has the same length as 'self'.
 
         method : {'backfill', 'bfill', 'pad', 'ffill', None}, default None
@@ -1050,12 +1421,11 @@ class GeometryArray(ExtensionArray):
             backfill / bfill: use NEXT valid observation to fill gap
 
         limit : int, default None
-            If method is specified, this is the maximum number of consecutive
-            NaN values to forward/backward fill. In other words, if there is
-            a gap with more than this number of consecutive NaNs, it will only
-            be partially filled. If method is not specified, this is the
-            maximum number of entries along the entire axis where NaNs will be
-            filled.
+            The maximum number of entries where NA values will be filled.
+
+        copy : bool, default True
+            Whether to make a copy of the data before filling. If False, then
+            the original should be modified and no new memory should be allocated.
 
         Returns
         -------
@@ -1065,10 +1435,36 @@ class GeometryArray(ExtensionArray):
             raise NotImplementedError("fillna with a method is not yet supported")
 
         mask = self.isna()
-        new_values = self.copy()
-        return new_values._fill(mask, value) if mask.any() else new_values
+        if copy:
+            new_values = self.copy()
+        else:
+            new_values = self
 
-    def astype(self, dtype, copy=True):
+        if not mask.any():
+            return new_values
+
+        if limit is not None and limit < len(self):
+            modify = mask.cumsum() > limit
+            if modify.any():
+                mask[modify] = False
+
+        if isna(value):
+            value = [None]
+        elif _is_scalar_geometry(value):
+            value = [value]
+        elif isinstance(value, GeometryArray):
+            value = value[mask]
+        else:
+            raise TypeError(
+                "'value' parameter must be None, a scalar geometry, or a GeoSeries, "
+                f"but you passed a {type(value).__name__!r}"
+            )
+        value_arr = np.asarray(value, dtype=object)
+
+        new_values._data[mask] = value_arr
+        return new_values
+
+    def astype(self, dtype, copy: bool = True) -> np.ndarray:
         """
         Cast to a NumPy array with 'dtype'.
 
@@ -1101,23 +1497,22 @@ class GeometryArray(ExtensionArray):
                 return pd.array(string_values, dtype=pd_dtype)
             return string_values.astype(dtype, copy=False)
         else:
-            return np.array(self, dtype=dtype, copy=copy)
+            # numpy 2.0 makes copy=False case strict (errors if cannot avoid the copy)
+            # -> in that case use `np.asarray` as backwards compatible alternative
+            # for `copy=None` (when requiring numpy 2+, this can be cleaned up)
+            if not copy:
+                return np.asarray(self, dtype=dtype)
+            else:
+                return np.array(self, dtype=dtype, copy=copy)
 
-    def isna(self):
-        """
-        Boolean NumPy array indicating if each value is missing
-        """
-        if compat.USE_SHAPELY_20:
-            return shapely.is_missing(self._data)
-        elif compat.USE_PYGEOS:
-            return pygeos.is_missing(self._data)
-        else:
-            return np.array([g is None for g in self._data], dtype="bool")
+    def isna(self) -> np.ndarray:
+        """Boolean NumPy array indicating if each value is missing."""
+        return shapely.is_missing(self._data)
 
     def value_counts(
         self,
         dropna: bool = True,
-    ):
+    ) -> pd.Series:
         """
         Compute a histogram of the counts of non-null values.
 
@@ -1130,11 +1525,10 @@ class GeometryArray(ExtensionArray):
         -------
         pd.Series
         """
-
         # note ExtensionArray usage of value_counts only specifies dropna,
         # so sort, normalize and bins are not arguments
         values = to_wkb(self)
-        from pandas import Series, Index
+        from pandas import Index, Series
 
         result = Series(values).value_counts(dropna=dropna)
         # value_counts converts None to nan, need to convert back for from_wkb to work
@@ -1142,11 +1536,11 @@ class GeometryArray(ExtensionArray):
         # Can't use fillna(None) or Index.putmask, as this gets converted back to nan
         # for object dtypes
         result.index = Index(
-            from_wkb(np.where(result.index.isna(), None, result.index))
+            from_wkb(np.where(result.index.isna(), None, result.index), crs=self.crs)
         )
         return result
 
-    def unique(self):
+    def unique(self) -> ExtensionArray:
         """Compute the ExtensionArray of unique values.
 
         Returns
@@ -1162,7 +1556,7 @@ class GeometryArray(ExtensionArray):
     def nbytes(self):
         return self._data.nbytes + len(self) * compat.GEOMETRY_OBJECT_SIZE
 
-    def shift(self, periods=1, fill_value=None):
+    def shift(self, periods: int = 1, fill_value: Any | None = None) -> GeometryArray:
         """
         Shift values by desired number.
 
@@ -1202,7 +1596,7 @@ class GeometryArray(ExtensionArray):
     # -------------------------------------------------------------------------
 
     @classmethod
-    def _from_sequence(cls, scalars, dtype=None, copy=False):
+    def _from_sequence(cls, scalars, dtype=None, copy: bool = False) -> ExtensionArray:
         """
         Construct a new ExtensionArray from a sequence of scalars.
 
@@ -1226,8 +1620,30 @@ class GeometryArray(ExtensionArray):
             scalars = [scalars]
         return from_shapely(scalars)
 
-    def _values_for_factorize(self):
-        # type: () -> Tuple[np.ndarray, Any]
+    @classmethod
+    def _from_sequence_of_strings(cls, strings, *, dtype=None, copy=False):
+        """
+        Construct a new ExtensionArray from a sequence of strings.
+
+        Parameters
+        ----------
+        strings : Sequence
+            Each element will be an instance of the scalar type for this
+            array, ``cls.dtype.type``.
+        dtype : dtype, optional
+            Construct for this particular dtype. This should be a Dtype
+            compatible with the ExtensionArray.
+        copy : bool, default False
+            If True, copy the underlying data.
+
+        Returns
+        -------
+        ExtensionArray
+        """
+        # GH 3099
+        return from_wkt(strings)
+
+    def _values_for_factorize(self) -> tuple[np.ndarray, Any]:
         """Return an array and missing value suitable for factorization.
 
         Returns
@@ -1246,7 +1662,7 @@ class GeometryArray(ExtensionArray):
         return vals, None
 
     @classmethod
-    def _from_factorized(cls, values, original):
+    def _from_factorized(cls, values: np.ndarray, original: ExtensionArray):
         """
         Reconstruct an ExtensionArray after factorization.
 
@@ -1264,8 +1680,7 @@ class GeometryArray(ExtensionArray):
         """
         return from_wkb(values, crs=original.crs)
 
-    def _values_for_argsort(self):
-        # type: () -> np.ndarray
+    def _values_for_argsort(self) -> np.ndarray:
         """Return values for sorting.
 
         Returns
@@ -1294,7 +1709,7 @@ class GeometryArray(ExtensionArray):
             # process those. The missing values are handled separately by
             # pandas regardless of the values we return here (to sort
             # first/last depending on 'na_position'), the distances for the
-            # empty geometries are substitued below with an appropriate value
+            # empty geometries are replaced below with an appropriate value
             geoms = self.copy()
             indices = np.nonzero(~mask)[0]
             if indices.size:
@@ -1329,8 +1744,8 @@ class GeometryArray(ExtensionArray):
     def argmax(self, skipna: bool = True) -> int:
         raise TypeError("geometries have no minimum or maximum")
 
-    def _formatter(self, boxed=False):
-        """Formatting function for scalar values.
+    def _formatter(self, boxed: bool = False):
+        """Return a formatting function for scalar values.
 
         This is used in the default '__repr__'. The returned formatting
         function receives instances of your scalar type.
@@ -1381,13 +1796,12 @@ class GeometryArray(ExtensionArray):
                         # typically projected coordinates
                         # (in case of unit meter: mm precision)
                         precision = 3
-            return lambda geom: shapely.wkt.dumps(geom, rounding_precision=precision)
+            return lambda geom: shapely.to_wkt(geom, rounding_precision=precision)
         return repr
 
     @classmethod
     def _concat_same_type(cls, to_concat):
-        """
-        Concatenate multiple array
+        """Concatenate multiple array.
 
         Parameters
         ----------
@@ -1400,26 +1814,28 @@ class GeometryArray(ExtensionArray):
         data = np.concatenate([ga._data for ga in to_concat])
         return GeometryArray(data, crs=_get_common_crs(to_concat))
 
-    def _reduce(self, name, skipna=True, **kwargs):
+    def _reduce(self, name: str, skipna: bool = True, keepdims: bool = False, **kwargs):
         # including the base class version here (that raises by default)
         # because this was not yet defined in pandas 0.23
-        if name == "any" or name == "all":
-            # TODO(pygeos)
-            return getattr(to_shapely(self), name)()
+        if name in ("any", "all"):
+            return getattr(self._data, name)(keepdims=keepdims)
         raise TypeError(
             f"'{type(self).__name__}' with dtype {self.dtype} "
             f"does not support reduction '{name}'"
         )
 
-    def __array__(self, dtype=None):
-        """
-        The numpy array interface.
+    def __array__(self, dtype=None, copy=None) -> np.ndarray:
+        """Return the data as a numpy array.
+
+        This is the numpy array interface.
 
         Returns
         -------
         values : numpy array
         """
-        return to_shapely(self)
+        if copy and (dtype is None or dtype == np.dtype("object")):
+            return self._data.copy()
+        return self._data
 
     def _binop(self, other, op):
         def convert_values(param):
@@ -1431,7 +1847,7 @@ class GeometryArray(ExtensionArray):
                 ovalues = [param] * len(self)
             return ovalues
 
-        if isinstance(other, (pd.Series, pd.Index, pd.DataFrame)):
+        if isinstance(other, pd.Series | pd.Index | pd.DataFrame):
             # rely on pandas to unbox and dispatch to us
             return NotImplemented
 
@@ -1451,14 +1867,18 @@ class GeometryArray(ExtensionArray):
     def __eq__(self, other):
         return self._binop(other, operator.eq)
 
+    # https://github.com/python/typeshed/issues/2148#issuecomment-520783318
+    # Incompatible types in assignment (expression has type "None", base class
+    # "object" defined the type as "Callable[[object], int]")
+    # (Explicitly mirrored from pandas to declare non hashable)
+    __hash__: ClassVar[None]  # type: ignore[assignment]
+
     def __ne__(self, other):
         return self._binop(other, operator.ne)
 
-    def __contains__(self, item):
-        """
-        Return for `item in self`.
-        """
-        if vectorized.isna(item):
+    def __contains__(self, item) -> bool:
+        """Return for `item in self`."""
+        if isna(item):
             if (
                 item is self.dtype.na_value
                 or isinstance(item, self.dtype.type)
@@ -1470,23 +1890,29 @@ class GeometryArray(ExtensionArray):
         return (self == item).any()
 
 
-def _get_common_crs(arr_seq):
+def _get_common_crs(arr_seq) -> CRS:
     # mask out all None arrays with no crs (most likely auto generated by pandas
     # from concat with missing column)
     arr_seq = [ga for ga in arr_seq if not (ga.isna().all() and ga.crs is None)]
+    # determine unique crs without using a set, because CRS hash can be different
+    # for objects with the same CRS
+    unique_crs = []
+    for arr in arr_seq:
+        if arr.crs not in unique_crs:
+            unique_crs.append(arr.crs)
 
-    crs_set = {arr.crs for arr in arr_seq}
-    crs_not_none = [crs for crs in crs_set if crs is not None]
+    crs_not_none = [crs for crs in unique_crs if crs is not None]
     names = [crs.name for crs in crs_not_none]
 
     if len(crs_not_none) == 0:
         return None
     if len(crs_not_none) == 1:
-        if len(crs_set) != 1:
+        if len(unique_crs) != 1:
             warnings.warn(
                 "CRS not set for some of the concatenation inputs. "
                 f"Setting output's CRS as {names[0]} "
-                "(the single non-null crs provided)."
+                "(the single non-null crs provided).",
+                stacklevel=2,
             )
         return crs_not_none[0]
 
@@ -1494,3 +1920,23 @@ def _get_common_crs(arr_seq):
         f"Cannot determine common CRS for concatenation inputs, got {names}. "
         "Use `to_crs()` to transform geometries to the same CRS before merging."
     )
+
+
+def transform(data, func) -> np.ndarray:
+    has_z = shapely.has_z(data)
+
+    result = np.empty_like(data)
+
+    coords = shapely.get_coordinates(data[~has_z], include_z=False)
+    new_coords_z = func(coords[:, 0], coords[:, 1])
+    result[~has_z] = shapely.set_coordinates(
+        data[~has_z].copy(), np.array(new_coords_z).T
+    )
+
+    coords_z = shapely.get_coordinates(data[has_z], include_z=True)
+    new_coords_z = func(coords_z[:, 0], coords_z[:, 1], coords_z[:, 2])
+    result[has_z] = shapely.set_coordinates(
+        data[has_z].copy(), np.array(new_coords_z).T
+    )
+
+    return result
