@@ -3,13 +3,14 @@ import pandas as pd
 from shapely.geometry import Point
 
 from geopandas import GeoDataFrame, GeoSeries
+from geopandas._compat import HAS_PYPROJ
 from geopandas.tools import geocode, reverse_geocode
 from geopandas.tools.geocoding import _prepare_geocode_result
 
+import pytest
+from geopandas.testing import assert_geodataframe_equal
 from geopandas.tests.util import assert_geoseries_equal, mock
 from pandas.testing import assert_series_equal
-from geopandas.testing import assert_geodataframe_equal
-import pytest
 
 geopy = pytest.importorskip("geopy")
 
@@ -45,7 +46,7 @@ class ReverseMock(mock.MagicMock):
         self._n = 0
 
     def __call__(self, *args, **kwargs):
-        self.return_value = "address{0}".format(self._n), args[0]
+        self.return_value = f"address{self._n}", args[0]
         self._n += 1
         return super().__call__(*args, **kwargs)
 
@@ -71,7 +72,8 @@ def test_prepare_result():
 
     df = _prepare_geocode_result(d)
     assert type(df) is GeoDataFrame
-    assert df.crs == "EPSG:4326"
+    if HAS_PYPROJ:
+        assert df.crs == "EPSG:4326"
     assert len(df) == 2
     assert "address" in df
 
@@ -93,20 +95,17 @@ def test_prepare_result_none():
 
     df = _prepare_geocode_result(d)
     assert type(df) is GeoDataFrame
-    assert df.crs == "EPSG:4326"
+    if HAS_PYPROJ:
+        assert df.crs == "EPSG:4326"
     assert len(df) == 2
     assert "address" in df
 
     row = df.loc["b"]
-    # The shapely.geometry.Point() is actually a GeometryCollection, and thus
-    # gets converted to that in conversion to pygeos. When converting back
-    # on access, you now get a GeometryCollection object instead of Point,
-    # which has no coords
-    # see https://github.com/Toblerity/Shapely/issues/742/#issuecomment-545296708
+
     # TODO we should probably replace this with a missing value instead of point?
-    # assert len(row["geometry"].coords) == 0
+    assert len(row["geometry"].coords) == 0
     assert row["geometry"].is_empty
-    assert row["address"] is None
+    assert pd.isna(row["address"])
 
 
 @pytest.mark.parametrize("geocode_result", (None, (None, None)))
