@@ -4,7 +4,7 @@ from pandas import Series
 
 from geopandas import GeoDataFrame
 
-from geopandas.testing import (  # noqa
+from geopandas.testing import (  # noqa: F401
     assert_geoseries_equal,
     geom_almost_equals,
     geom_equals,
@@ -13,12 +13,21 @@ from geopandas.testing import (  # noqa
 HERE = os.path.abspath(os.path.dirname(__file__))
 PACKAGE_DIR = os.path.dirname(os.path.dirname(HERE))
 
+_TEST_DATA_DIR = os.path.join(PACKAGE_DIR, "geopandas", "tests", "data")
+_NYBB = "zip://" + os.path.join(_TEST_DATA_DIR, "nybb_16a.zip")
+_NATURALEARTH_CITIES = os.path.join(
+    _TEST_DATA_DIR, "naturalearth_cities", "naturalearth_cities.shp"
+)
+_NATURALEARTH_LOWRES = os.path.join(
+    _TEST_DATA_DIR, "naturalearth_lowres", "naturalearth_lowres.shp"
+)
+
 
 # mock not used here, but the import from here is used in other modules
 try:
-    import unittest.mock as mock  # noqa
+    from unittest import mock
 except ImportError:
-    import mock  # noqa
+    from unittest import mock  # noqa: F401
 
 
 def validate_boro_df(df, case_sensitive=False):
@@ -91,7 +100,7 @@ def create_spatialite(con, df):
 
 def create_postgis(con, df, srid=None, geom_col="geom"):
     """
-    Create a nybb table in the test_geopandas PostGIS database.
+    Create an nybb table in the test_geopandas PostGIS database.
     Returns a boolean indicating whether the database table was successfully
     created
     """
@@ -102,8 +111,8 @@ def create_postgis(con, df, srid=None, geom_col="geom"):
     # > createdb test_geopandas
     # > psql -c "CREATE EXTENSION postgis" -d test_geopandas
     if srid is not None:
-        geom_schema = "geometry(MULTIPOLYGON, {})".format(srid)
-        geom_insert = "ST_SetSRID(ST_GeometryFromText(%s), {})".format(srid)
+        geom_schema = f"geometry(MULTIPOLYGON, {srid})"
+        geom_insert = f"ST_SetSRID(ST_GeometryFromText(%s), {srid})"
     else:
         geom_schema = "geometry"
         geom_insert = "ST_GeometryFromText(%s)"
@@ -111,22 +120,18 @@ def create_postgis(con, df, srid=None, geom_col="geom"):
         cursor = con.cursor()
         cursor.execute("DROP TABLE IF EXISTS nybb;")
 
-        sql = """CREATE TABLE nybb (
+        sql = f"""CREATE TABLE nybb (
             {geom_col}   {geom_schema},
             borocode     integer,
             boroname     varchar(40),
             shape_leng   float,
             shape_area   float
-            );""".format(
-            geom_col=geom_col, geom_schema=geom_schema
-        )
+            );"""
         cursor.execute(sql)
 
         for i, row in df.iterrows():
-            sql = """INSERT INTO nybb VALUES ({}, %s, %s, %s, %s
-            );""".format(
-                geom_insert
-            )
+            sql = f"""INSERT INTO nybb VALUES ({geom_insert}, %s, %s, %s, %s
+            );"""
             cursor.execute(
                 sql,
                 (
