@@ -4,7 +4,7 @@ import sys
 
 
 def _get_sys_info():
-    """System information
+    """System information.
 
     Returns
     -------
@@ -23,7 +23,8 @@ def _get_sys_info():
 
 
 def _get_C_info():
-    """Information on system PROJ, GDAL, GEOS
+    """Information on system PROJ, GDAL, GEOS.
+
     Returns
     -------
     c_info: dict
@@ -58,33 +59,27 @@ def _get_C_info():
             geos_dir = None
 
     try:
-        import fiona
+        import pyogrio
 
-        gdal_version = fiona.env.get_gdal_release_name()
+        gdal_version = pyogrio.__gdal_version_string__
+        gdal_dir = pyogrio.get_gdal_data_path()
     except Exception:
         gdal_version = None
-    try:
-        import fiona
-
-        gdal_dir = fiona.env.GDALDataFinder().search()
-    except Exception:
         gdal_dir = None
 
     if gdal_version is None:
         try:
-            import pyogrio
+            import fiona
 
-            gdal_version = pyogrio.__gdal_version_string__
-            gdal_dir = None
+            gdal_version = fiona.env.get_gdal_release_name()
         except Exception:
-            pass
+            gdal_version = None
         try:
-            # get_gdal_data_path is only available in pyogrio >= 0.4.2
-            from pyogrio import get_gdal_data_path
+            import fiona
 
-            gdal_dir = get_gdal_data_path()
+            gdal_dir = fiona.env.GDALDataFinder().search()
         except Exception:
-            pass
+            gdal_dir = None
 
     blob = [
         ("GEOS", geos_version),
@@ -99,7 +94,7 @@ def _get_C_info():
 
 
 def _get_deps_info():
-    """Overview of the installed version of main dependencies
+    """Overview of the installed version of main dependencies.
 
     Returns
     -------
@@ -114,16 +109,15 @@ def _get_deps_info():
         "pyproj",
         "shapely",
         # optional deps
-        "fiona",
+        "pyogrio",
         "geoalchemy2",
         "geopy",
         "matplotlib",
         "mapclassify",
-        "pygeos",
-        "pyogrio",
+        "fiona",
+        "psycopg",
         "psycopg2",
         "pyarrow",
-        "rtree",
     ]
 
     def get_version(module):
@@ -151,7 +145,6 @@ def show_versions():
 
     Examples
     --------
-
     ::
 
         $ python -c "import geopandas; geopandas.show_versions()"
@@ -161,7 +154,7 @@ def show_versions():
     proj_info = _get_C_info()
 
     maxlen = max(len(x) for x in deps_info)
-    tpl = "{{k:<{maxlen}}}: {{stat}}".format(maxlen=maxlen)
+    tpl = f"{{k:<{maxlen}}}: {{stat}}"
     print("\nSYSTEM INFO")
     print("-----------")
     for k, stat in sys_info.items():
