@@ -1,28 +1,27 @@
-from geopandas._config import options  # noqa
+from geopandas._config import options
 
-from geopandas.geoseries import GeoSeries  # noqa
-from geopandas.geodataframe import GeoDataFrame  # noqa
-from geopandas.array import points_from_xy  # noqa
+from geopandas.geoseries import GeoSeries
+from geopandas.geodataframe import GeoDataFrame
+from geopandas.array import points_from_xy
 
-from geopandas.io.file import _read_file as read_file  # noqa
-from geopandas.io.arrow import _read_parquet as read_parquet  # noqa
-from geopandas.io.arrow import _read_feather as read_feather  # noqa
-from geopandas.io.sql import _read_postgis as read_postgis  # noqa
-from geopandas.tools import sjoin, sjoin_nearest  # noqa
-from geopandas.tools import overlay  # noqa
-from geopandas.tools._show_versions import show_versions  # noqa
-from geopandas.tools import clip  # noqa
+from geopandas.io.file import _read_file as read_file
+from geopandas.io.file import _list_layers as list_layers
+from geopandas.io.arrow import _read_parquet as read_parquet
+from geopandas.io.arrow import _read_feather as read_feather
+from geopandas.io.sql import _read_postgis as read_postgis
+from geopandas.tools import sjoin, sjoin_nearest
+from geopandas.tools import overlay
+from geopandas.tools._show_versions import show_versions
+from geopandas.tools import clip
 
 
-import geopandas.datasets  # noqa
+import geopandas.datasets
 
 
 # make the interactive namespace easier to use
 # for `from geopandas import *` demos.
-import geopandas as gpd  # noqa
-import pandas as pd  # noqa
-import numpy as np  # noqa
+import geopandas as gpd
+import pandas as pd
+import numpy as np
 
-from . import _version
-
-__version__ = _version.get_versions()["version"]
+from geopandas._version import __version__
