@@ -1,40 +1,49 @@
+from __future__ import annotations
+
 import warnings
+from collections.abc import Collection, Iterable, Sequence
+from typing import TYPE_CHECKING, Any, Literal
 
 import numpy as np
 import pandas as pd
+from pandas import CategoricalDtype
 from pandas.plotting import PlotAccessor
+
+import shapely
 
 import geopandas
 
-from packaging.version import Version
-
 from ._decorator import doc
 
-
-def deprecated(new, warning_type=FutureWarning):
-    """Helper to provide deprecation warning."""
-
-    def old(*args, **kwargs):
-        warnings.warn(
-            "{} is intended for internal ".format(new.__name__[1:])
-            + "use only, and will be deprecated.",
-            warning_type,
-            stacklevel=2,
-        )
-        new(*args, **kwargs)
-
-    return old
+if TYPE_CHECKING:
+    from matplotlib.axes import Axes
+    from matplotlib.collections import (
+        LineCollection,
+        PatchCollection,
+        PathCollection,
+    )
+    from matplotlib.colors import Colormap
+    from matplotlib.markers import MarkerStyle
+    from matplotlib.patches import PathPatch
+    from matplotlib.path import Path
 
 
-def _sanitize_geoms(geoms, prefix="Multi"):
-    """
-    Returns Series like geoms and index, except that any Multi geometries
+def _sanitize_geoms(
+    geoms: geopandas.GeoSeries, prefix: str = "Multi"
+) -> tuple[geopandas.GeoSeries, np.ndarray]:
+    """Return sanitized geometry with the indices of original geometry.
+
+    1. Normalize all geometry to ensure holes are correctly plotted.
+    2. Explode multi-part geometries to individual components. This generates an
+       index where values are repeated for all components in the same
+       collection.
+    3. Filter out missing and empty geometry. The resulting index does not contain
+       their IDs.
+
+    Series like geoms and index, except that any Multi geometries
     are split into their components and indices are repeated for all component
     in the same Multi geometry. At the same time, empty or missing geometries are
-    filtered out.  Maintains 1:1 matching of geometry to value.
-
-    Prefix specifies type of geometry to be flatten. 'Multi' for MultiPoint and similar,
-    "Geom" for GeometryCollection.
+    filtered out. The index then maintains 1:1 matching of geometry to value.
 
     Returns
     -------
@@ -45,6 +54,7 @@ def _sanitize_geoms(geoms, prefix="Multi"):
     """
     # TODO(shapely) look into simplifying this with
     # shapely.get_parts(geoms, return_index=True) from shapely 2.0
+    geoms = geoms.normalize()
     components, component_index = [], []
 
     if (
@@ -65,27 +75,19 @@ def _sanitize_geoms(geoms, prefix="Multi"):
             components.append(geom)
             component_index.append(ix)
 
-    return components, np.array(component_index)
+    return geopandas.GeoSeries(components, crs=geoms.crs), np.array(component_index)
 
 
-def _expand_kwargs(kwargs, multiindex):
+def _expand_kwargs(kwargs: dict, multiindex: np.ndarray) -> None:
     """
     Most arguments to the plot functions must be a (single) value, or a sequence
     of values. This function checks each key-value pair in 'kwargs' and expands
     it (in place) to the correct length/formats with help of 'multiindex', unless
     the value appears to already be a valid (single) value for the key.
     """
-    import matplotlib
     from matplotlib.colors import is_color_like
-    from typing import Iterable
 
-    mpl = Version(matplotlib.__version__)
-    if mpl >= Version("3.4"):
-        # alpha is supported as array argument with matplotlib 3.4+
-        scalar_kwargs = ["marker", "path_effects"]
-    else:
-        scalar_kwargs = ["marker", "alpha", "path_effects"]
-
+    scalar_kwargs = ["marker", "path_effects"]
     for att, value in kwargs.items():
         if "color" in att:  # color(s), edgecolor(s), facecolor(s)
             if is_color_like(value):
@@ -106,11 +108,15 @@ def _expand_kwargs(kwargs, multiindex):
             kwargs[att] = np.take(value, multiindex, axis=0)
 
 
-def _PolygonPatch(polygon, **kwargs):
-    """Constructs a matplotlib patch from a Polygon geometry
+def _PolygonPatch(polygon: shapely.Geometry, **kwargs) -> PathPatch:
+    """Construct a matplotlib patch from a (Multi)Polygon geometry.
 
     The `kwargs` are those supported by the matplotlib.patches.PathPatch class
     constructor. Returns an instance of matplotlib.patches.PathPatch.
+
+    To ensure proper rendering on the matplotlib side, winding order of individual
+    rings needs to be normalized as the order is what matplotlib uses to determine
+    if a Path represents a patch or a hole.
 
     Example (using Shapely Point and a matplotlib axes)::
 
@@ -125,102 +131,144 @@ def _PolygonPatch(polygon, **kwargs):
     from matplotlib.patches import PathPatch
     from matplotlib.path import Path
 
-    path = Path.make_compound_path(
-        Path(np.asarray(polygon.exterior.coords)[:, :2]),
-        *[Path(np.asarray(ring.coords)[:, :2]) for ring in polygon.interiors],
-    )
+    if polygon.geom_type == "Polygon":
+        parts = [polygon]
+    else:
+        parts = polygon.geoms
+    paths = []
+    for part in parts:
+        # exteriors
+        paths.append(Path(np.asarray(part.exterior.coords)[:, :2], closed=True))
+        # interiors
+        paths.extend(
+            Path(np.asarray(ring.coords)[:, :2], closed=True) for ring in part.interiors
+        )
+    path = Path.make_compound_path(*paths)
+
     return PathPatch(path, **kwargs)
 
 
 def _plot_polygon_collection(
-    ax, geoms, values=None, color=None, cmap=None, vmin=None, vmax=None, **kwargs
-):
-    """
-    Plots a collection of Polygon and MultiPolygon geometries to `ax`
+    ax: Axes,
+    geoms: geopandas.GeoSeries,
+    values: np.ndarray | None = None,
+    cmap: str | Colormap | None = None,
+    vmin: float | int | None = None,
+    vmax: float | int | None = None,
+    autolim: bool = True,
+    **kwargs,
+) -> PatchCollection:
+    """Plot a collection of Polygon and MultiPolygon geometries to `ax`.
+
+    Note that all style keywords, like ``color`` that can be set as an array in
+    matplotlib shall be passed directly via kwargs.
+
+    No need to explode geometries to single-parts as _PolygonPatch supports
+    MultiPolygons,
 
     Parameters
     ----------
     ax : matplotlib.axes.Axes
-        where shapes will be plotted
-    geoms : a sequence of `N` Polygons and/or MultiPolygons (can be mixed)
-
-    values : a sequence of `N` values, optional
-        Values will be mapped to colors using vmin/vmax/cmap. They should
+        Axes on which to add the collection
+    geoms : GeoSeries
+        GeoSeries of (Multi)Polygons
+    values : np.ndarray, optional
+        Values will be mapped to colors using vmin/vmax/norm/cmap. They should
         have 1:1 correspondence with the geometries (not their components).
-        Otherwise follows `color` / `facecolor` kwargs.
-    edgecolor : single color or sequence of `N` colors
-        Color for the edge of the polygons
-    facecolor : single color or sequence of `N` colors
-        Color to fill the polygons. Cannot be used together with `values`.
-    color : single color or sequence of `N` colors
-        Sets both `edgecolor` and `facecolor`
-    **kwargs
-        Additional keyword arguments passed to the collection
+        Otherwise follows `color` / `facecolor` kwargs, by default None
+    cmap : str or Colormap, optional
+        The colormap recognized by matplotlib, by default None
+    vmin : float, optional
+        Minimum value of cmap, by default None
+    vmax : float, optional
+        Maximum value of cmap, by default None
+    autolim : bool, optional
+        Update axes data limits to contain the new geometries, by default True
 
     Returns
     -------
-    collection : matplotlib.collections.Collection that was plotted
+    _GeoPandasPolyCollection
+        matplotlib.collections.Collection that was plotted
     """
     from matplotlib.collections import PatchCollection
+    from matplotlib.legend import Legend
+    from matplotlib.legend_handler import HandlerPolyCollection
 
-    geoms, multiindex = _sanitize_geoms(geoms)
-    if values is not None:
-        values = np.take(values, multiindex, axis=0)
+    class _GeoPandasPolyCollection(PatchCollection):
+        """Subclass to assign handler without overriding one for PatchCollection."""
 
-    # PatchCollection does not accept some kwargs.
+    # PatchCollection is not supported by Legend but we can use PolyCollection handler
+    # instead in our specific case. Define a subclass and assign a handler.
+    Legend.update_default_handler_map(
+        {_GeoPandasPolyCollection: HandlerPolyCollection()}
+    )
+
+    # _GeoPandasPolyCollection does not accept some kwargs.
     kwargs = {
         att: value
         for att, value in kwargs.items()
         if att not in ["markersize", "marker"]
     }
 
-    # Add to kwargs for easier checking below.
-    if color is not None:
-        kwargs["color"] = color
-
-    _expand_kwargs(kwargs, multiindex)
-
-    collection = PatchCollection([_PolygonPatch(poly) for poly in geoms], **kwargs)
+    collection = _GeoPandasPolyCollection(
+        [_PolygonPatch(poly) for poly in geoms], **kwargs
+    )
 
     if values is not None:
         collection.set_array(np.asarray(values))
-        collection.set_cmap(cmap)
+        if cmap:
+            collection.set_cmap(cmap)
         if "norm" not in kwargs:
             collection.set_clim(vmin, vmax)
 
-    ax.add_collection(collection, autolim=True)
+    ax.add_collection(collection, autolim=autolim)
     ax.autoscale_view()
     return collection
 
 
-plot_polygon_collection = deprecated(_plot_polygon_collection)
-
-
 def _plot_linestring_collection(
-    ax, geoms, values=None, color=None, cmap=None, vmin=None, vmax=None, **kwargs
-):
-    """
-    Plots a collection of LineString and MultiLineString geometries to `ax`
+    ax: Axes,
+    geoms: geopandas.GeoSeries,
+    values: np.ndarray | None = None,
+    color: str | Sequence | None = None,
+    cmap: str | Colormap | None = None,
+    vmin: float | int | None = None,
+    vmax: float | int | None = None,
+    autolim: bool = True,
+    **kwargs,
+) -> LineCollection:
+    """Plot a collection of LineString and MultiLineString geometries to `ax`.
 
     Parameters
     ----------
     ax : matplotlib.axes.Axes
-        where shapes will be plotted
-    geoms : a sequence of `N` LineStrings and/or MultiLineStrings (can be
-            mixed)
-    values : a sequence of `N` values, optional
-        Values will be mapped to colors using vmin/vmax/cmap. They should
+        Axes on which to add the collection
+    geoms : GeoSeries
+        GeoSeries of (Multi)Polygons
+    values : np.ndarray, optional
+        Values will be mapped to colors using vmin/vmax/norm/cmap. They should
         have 1:1 correspondence with the geometries (not their components).
+        Otherwise follows `color` / `facecolor` kwargs, by default None
     color : single color or sequence of `N` colors
-        Cannot be used together with `values`.
+        Color definition understood by matplotlib. Cannot be used together with
+        `values`.
+    cmap : str or Colormap, optional
+        The colormap recognized by matplotlib, by default None
+    vmin : float, optional
+        Minimum value of cmap, by default None
+    vmax : float, optional
+        Maximum value of cmap, by default None
+    autolim : bool, optional
+        Update axes data limits to contain the new geometries, by default True
 
     Returns
     -------
-    collection : matplotlib.collections.Collection that was plotted
+    LineCollection
+        matplotlib.collections.Collection that was plotted
     """
     from matplotlib.collections import LineCollection
 
-    geoms, multiindex = _sanitize_geoms(geoms)
+    geoms, multiindex = shapely.get_parts(geoms.values, return_index=True)
     if values is not None:
         values = np.take(values, multiindex, axis=0)
 
@@ -242,42 +290,51 @@ def _plot_linestring_collection(
 
     if values is not None:
         collection.set_array(np.asarray(values))
-        collection.set_cmap(cmap)
+        if cmap:
+            collection.set_cmap(cmap)
         if "norm" not in kwargs:
             collection.set_clim(vmin, vmax)
 
-    ax.add_collection(collection, autolim=True)
+    ax.add_collection(collection, autolim=autolim)
     ax.autoscale_view()
     return collection
 
 
-plot_linestring_collection = deprecated(_plot_linestring_collection)
-
-
 def _plot_point_collection(
-    ax,
-    geoms,
-    values=None,
-    color=None,
-    cmap=None,
-    vmin=None,
-    vmax=None,
-    marker="o",
-    markersize=None,
+    ax: Axes,
+    geoms: geopandas.GeoSeries,
+    values: np.ndarray | None = None,
+    color: str | Sequence | None = None,
+    cmap: str | Colormap | None = None,
+    vmin: float | int | None = None,
+    vmax: float | int | None = None,
+    marker: str | MarkerStyle | Path = "o",
+    markersize: float | Sequence[float] | None = None,
     **kwargs,
-):
-    """
-    Plots a collection of Point and MultiPoint geometries to `ax`
+) -> PathCollection:
+    """Plot a collection of Point and MultiPoint geometries to `ax`.
 
     Parameters
     ----------
     ax : matplotlib.axes.Axes
-        where shapes will be plotted
-    geoms : sequence of `N` Points or MultiPoints
-
-    values : a sequence of `N` values, optional
-        Values mapped to colors using vmin, vmax, and cmap.
-        Cannot be specified together with `color`.
+        Axes on which to add the collection
+    geoms : GeoSeries
+        GeoSeries of (Multi)Polygons
+    values : np.ndarray, optional
+        Values will be mapped to colors using vmin/vmax/norm/cmap. They should
+        have 1:1 correspondence with the geometries (not their components).
+        Otherwise follows `color` / `facecolor` kwargs, by default None
+    color : single color or sequence of `N` colors
+        Color definition understood by matplotlib. Cannot be used together with
+        `values`.
+    cmap : str or Colormap, optional
+        The colormap recognized by matplotlib, by default None
+    vmin : float, optional
+        Minimum value of cmap, by default None
+    vmax : float, optional
+        Maximum value of cmap, by default None
+    marker : str, MarkerStyle, Path
+        Style of the marker to be used.
     markersize : scalar or array-like, optional
         Size of the markers. Note that under the hood ``scatter`` is
         used, so the specified value will be proportional to the
@@ -287,42 +344,43 @@ def _plot_point_collection(
     -------
     collection : matplotlib.collections.Collection that was plotted
     """
-    if values is not None and color is not None:
-        raise ValueError("Can only specify one of 'values' and 'color' kwargs")
+    geoms, multiindex = shapely.get_parts(geoms.values, return_index=True)
 
-    geoms, multiindex = _sanitize_geoms(geoms)
-    # values are expanded below as kwargs["c"]
+    xy = shapely.get_coordinates(geoms)
 
-    x = [p.x if not p.is_empty else None for p in geoms]
-    y = [p.y if not p.is_empty else None for p in geoms]
-
-    # matplotlib 1.4 does not support c=None, and < 2.0 does not support s=None
+    # Add to kwargs for easier checking below.
     if values is not None:
         kwargs["c"] = values
     if markersize is not None:
         kwargs["s"] = markersize
-
-    # Add to kwargs for easier checking below.
     if color is not None:
         kwargs["color"] = color
     if marker is not None:
         kwargs["marker"] = marker
+
     _expand_kwargs(kwargs, multiindex)
 
+    # norm cannot be passed alongside vmin and vmax
     if "norm" not in kwargs:
-        collection = ax.scatter(x, y, vmin=vmin, vmax=vmax, cmap=cmap, **kwargs)
+        collection = ax.scatter(
+            xy[:, 0], xy[:, 1], vmin=vmin, vmax=vmax, cmap=cmap, **kwargs
+        )
     else:
-        collection = ax.scatter(x, y, cmap=cmap, **kwargs)
+        collection = ax.scatter(xy[:, 0], xy[:, 1], cmap=cmap, **kwargs)
 
     return collection
 
 
-plot_point_collection = deprecated(_plot_point_collection)
-
-
 def plot_series(
-    s, cmap=None, color=None, ax=None, figsize=None, aspect="auto", **style_kwds
-):
+    s: geopandas.GeoSeries,
+    cmap: str | Colormap | None = None,
+    color: str | Sequence | None = None,
+    ax: Axes | None = None,
+    figsize: tuple[float, float] | None = None,
+    aspect: float | Literal["auto", "equal", None] = "auto",
+    autolim: bool = True,
+    **style_kwds,
+) -> Axes:
     """
     Plot a GeoSeries.
 
@@ -357,6 +415,8 @@ def plot_series(
         square appears square in the middle of the plot. This implies an
         Equirectangular projection. If None, the aspect of `ax` won't be changed. It can
         also be set manually (float) as the ratio of y-unit to x-unit.
+    autolim : bool (default True)
+        Update axes data limits to contain the new geometries.
     **style_kwds : dict
         Color options to be passed on to the actual plot function, such
         as ``edgecolor``, ``facecolor``, ``linewidth``, ``markersize``,
@@ -366,33 +426,40 @@ def plot_series(
     -------
     ax : matplotlib axes instance
     """
-    if "colormap" in style_kwds:
-        warnings.warn(
-            "'colormap' is deprecated, please use 'cmap' instead "
-            "(for consistency with matplotlib)",
-            FutureWarning,
-        )
-        cmap = style_kwds.pop("colormap")
-    if "axes" in style_kwds:
-        warnings.warn(
-            "'axes' is deprecated, please use 'ax' instead "
-            "(for consistency with pandas)",
-            FutureWarning,
-        )
-        ax = style_kwds.pop("axes")
-
     try:
-        import matplotlib.pyplot as plt
-    except ImportError:
+        import matplotlib  # noqa: F401
+        from matplotlib.colors import Colormap
+    except ImportError as err:
         raise ImportError(
             "The matplotlib package is required for plotting in geopandas. "
             "You can install it using 'conda install -c conda-forge matplotlib' or "
             "'pip install matplotlib'."
-        )
+        ) from err
+
+    import matplotlib.pyplot as plt
 
     if ax is None:
-        fig, ax = plt.subplots(figsize=figsize)
+        _, ax = plt.subplots(figsize=figsize)
 
+    if s.empty:
+        warnings.warn(
+            "The GeoSeries you are attempting to plot is "
+            "empty. Nothing has been displayed.",
+            UserWarning,
+            stacklevel=3,
+        )
+        return ax
+
+    if s.is_empty.all():
+        warnings.warn(
+            "The GeoSeries you are attempting to plot is "
+            "composed of empty geometries. Nothing has been displayed.",
+            UserWarning,
+            stacklevel=3,
+        )
+        return ax
+
+    # set correct aspect to preserve proportions in geographic CRS
     if aspect == "auto":
         if s.crs and s.crs.is_geographic:
             bounds = s.total_bounds
@@ -405,44 +472,36 @@ def plot_series(
     elif aspect is not None:
         ax.set_aspect(aspect)
 
-    if s.empty:
-        warnings.warn(
-            "The GeoSeries you are attempting to plot is "
-            "empty. Nothing has been displayed.",
-            UserWarning,
-        )
-        return ax
+    # decompose GeometryCollections
+    geoms, multiindex = _sanitize_geoms(s, "Geom")
 
-    if s.is_empty.all():
-        warnings.warn(
-            "The GeoSeries you are attempting to plot is "
-            "composed of empty geometries. Nothing has been displayed.",
-            UserWarning,
-        )
-        return ax
-
-    # have colors been given for all geometries?
-    color_given = pd.api.types.is_list_like(color) and len(color) == len(s)
+    values = None
+    color_given = False
 
     # if cmap is specified, create range of colors based on cmap
-    values = None
     if cmap is not None:
         values = np.arange(len(s))
-        if hasattr(cmap, "N"):
+        if isinstance(cmap, Colormap) and hasattr(cmap, "N"):
+            # repeat for cmap with limited number of colors
             values = values % cmap.N
-        style_kwds["vmin"] = style_kwds.get("vmin", values.min())
-        style_kwds["vmax"] = style_kwds.get("vmax", values.max())
+        style_kwds["vmin"] = values.min()
+        style_kwds["vmax"] = values.max()
 
-    # decompose GeometryCollections
-    geoms, multiindex = _sanitize_geoms(s.geometry, prefix="Geom")
-    values = np.take(values, multiindex, axis=0) if cmap else None
-    # ensure indexes are consistent
-    if color_given and isinstance(color, pd.Series):
-        color = color.reindex(s.index)
-    expl_color = np.take(color, multiindex, axis=0) if color_given else color
-    expl_series = geopandas.GeoSeries(geoms)
+        # ensure proper mapping of values to components of GeometryCollections
+        values = np.take(values, multiindex, axis=0)
 
-    geom_types = expl_series.geom_type
+    # if color is specified as a list-like, ensure it is properly mapped to components
+    elif color is not None:
+        color_given = pd.api.types.is_list_like(color) and len(color) == len(s)
+        # have colors been given for all geometries?
+        if color_given:
+            # ensure indexes are consistent
+            if isinstance(color, pd.Series):
+                color = color.reindex(s.index)
+            color = np.take(color, multiindex, axis=0)
+
+    # subdivide by geometry type - each has its own collection
+    geom_types = geoms.geom_type
     poly_idx = np.asarray((geom_types == "Polygon") | (geom_types == "MultiPolygon"))
     line_idx = np.asarray(
         (geom_types == "LineString")
@@ -452,41 +511,51 @@ def plot_series(
     point_idx = np.asarray((geom_types == "Point") | (geom_types == "MultiPoint"))
 
     # plot all Polygons and all MultiPolygon components in the same collection
-    polys = expl_series[poly_idx]
+    polys = geoms[poly_idx]
     if not polys.empty:
         # color overrides both face and edgecolor. As we want people to be
         # able to use edgecolor as well, pass color to facecolor
         facecolor = style_kwds.pop("facecolor", None)
-        color_ = expl_color[poly_idx] if color_given else color
-        if color is not None:
-            facecolor = color_
 
-        values_ = values[poly_idx] if cmap else None
+        if color is not None:
+            facecolor = color[poly_idx] if color_given else color
+
+        values_ = values[poly_idx] if values is not None else None
+
         _plot_polygon_collection(
-            ax, polys, values_, facecolor=facecolor, cmap=cmap, **style_kwds
+            ax,
+            polys,
+            values_,
+            facecolor=facecolor,
+            cmap=cmap,
+            autolim=autolim,
+            **style_kwds,
         )
 
     # plot all LineStrings and MultiLineString components in same collection
-    lines = expl_series[line_idx]
+    lines = geoms[line_idx]
     if not lines.empty:
-        values_ = values[line_idx] if cmap else None
-        color_ = expl_color[line_idx] if color_given else color
+        values_ = values[line_idx] if values is not None else None
+
+        color_ = color[line_idx] if color_given else color
 
         _plot_linestring_collection(
-            ax, lines, values_, color=color_, cmap=cmap, **style_kwds
+            ax, lines, values_, color=color_, cmap=cmap, autolim=autolim, **style_kwds
         )
 
     # plot all Points in the same collection
-    points = expl_series[point_idx]
+    points = geoms[point_idx]
     if not points.empty:
-        values_ = values[point_idx] if cmap else None
-        color_ = expl_color[point_idx] if color_given else color
+        values_ = values[point_idx] if values is not None else None
+
+        color_ = color[point_idx] if color_given else color
 
         _plot_point_collection(
             ax, points, values_, color=color_, cmap=cmap, **style_kwds
         )
 
-    plt.draw()
+    ax.figure.canvas.draw_idle()
+
     return ax
 
 
@@ -510,6 +579,7 @@ def plot_dataframe(
     classification_kwds=None,
     missing_kwds=None,
     aspect="auto",
+    autolim=True,
     **style_kwds,
 ):
     """
@@ -521,11 +591,11 @@ def plot_dataframe(
 
     Parameters
     ----------
-    column : str, np.array, pd.Series (default None)
-        The name of the dataframe column, np.array, or pd.Series to be plotted.
-        If np.array or pd.Series are used then it must have same length as
-        dataframe. Values are used to color the plot. Ignored if `color` is
-        also set.
+    column : str, np.array, pd.Series, pd.Index (default None)
+        The name of the dataframe column, np.array, pd.Series, or pd.Index
+        to be plotted. If np.array, pd.Series, or pd.Index are used then it
+        must have same length as dataframe. Values are used to color the plot.
+        Ignored if `color` is also set.
     kind: str
         The kind of plots to produce. The default is to create a map ("geo").
         Other supported kinds of plots from pandas:
@@ -613,7 +683,8 @@ def plot_dataframe(
         square appears square in the middle of the plot. This implies an
         Equirectangular projection. If None, the aspect of `ax` won't be changed. It can
         also be set manually (float) as the ratio of y-unit to x-unit.
-
+    autolim : bool (default True)
+        Update axes data limits to contain the new geometries.
     **style_kwds : dict
         Style options to be passed on to the actual plot function, such
         as ``edgecolor``, ``facecolor``, ``linewidth``, ``markersize``,
@@ -625,44 +696,28 @@ def plot_dataframe(
 
     Examples
     --------
-    >>> df = geopandas.read_file(geopandas.datasets.get_path("naturalearth_lowres"))
+    >>> import geodatasets
+    >>> df = geopandas.read_file(geodatasets.get_path("nybb"))
     >>> df.head()  # doctest: +SKIP
-        pop_est      continent                      name iso_a3  \
-gdp_md_est                                           geometry
-    0     920938        Oceania                      Fiji    FJI      8374.0  MULTIPOLY\
-GON (((180.00000 -16.06713, 180.00000...
-    1   53950935         Africa                  Tanzania    TZA    150600.0  POLYGON (\
-(33.90371 -0.95000, 34.07262 -1.05982...
-    2     603253         Africa                 W. Sahara    ESH       906.5  POLYGON (\
-(-8.66559 27.65643, -8.66512 27.58948...
-    3   35623680  North America                    Canada    CAN   1674000.0  MULTIPOLY\
-GON (((-122.84000 49.00000, -122.9742...
-    4  326625791  North America  United States of America    USA  18560000.0  MULTIPOLY\
-GON (((-122.84000 49.00000, -120.0000...
+       BoroCode  ...                                           geometry
+    0         5  ...  MULTIPOLYGON (((970217.022 145643.332, 970227....
+    1         4  ...  MULTIPOLYGON (((1029606.077 156073.814, 102957...
+    2         3  ...  MULTIPOLYGON (((1021176.479 151374.797, 102100...
+    3         1  ...  MULTIPOLYGON (((981219.056 188655.316, 980940....
+    4         2  ...  MULTIPOLYGON (((1012821.806 229228.265, 101278...
 
-    >>> df.plot("pop_est", cmap="Blues")  # doctest: +SKIP
+    >>> df.plot("BoroName", cmap="Set1")  # doctest: +SKIP
 
     See the User Guide page :doc:`../../user_guide/mapping` for details.
 
     """
-    if "colormap" in style_kwds:
-        warnings.warn(
-            "'colormap' is deprecated, please use 'cmap' instead "
-            "(for consistency with matplotlib)",
-            FutureWarning,
-        )
-        cmap = style_kwds.pop("colormap")
-    if "axes" in style_kwds:
-        warnings.warn(
-            "'axes' is deprecated, please use 'ax' instead "
-            "(for consistency with pandas)",
-            FutureWarning,
-        )
-        ax = style_kwds.pop("axes")
     if column is not None and color is not None:
         warnings.warn(
-            "Only specify one of 'column' or 'color'. Using 'color'.", UserWarning
+            "Only specify one of 'column' or 'color'. Using 'color'.",
+            UserWarning,
+            stacklevel=3,
         )
+        # TODO should these both be supported if column is a column-name?
         column = None
 
     try:
@@ -677,7 +732,7 @@ GON (((-122.84000 49.00000, -120.0000...
     if ax is None:
         if cax is not None:
             raise ValueError("'ax' can not be None if 'cax' is not.")
-        fig, ax = plt.subplots(figsize=figsize)
+        _fig, ax = plt.subplots(figsize=figsize)
 
     if aspect == "auto":
         if df.crs and df.crs.is_geographic:
@@ -701,6 +756,7 @@ GON (((-122.84000 49.00000, -120.0000...
             "The GeoDataFrame you are attempting to plot is "
             "empty. Nothing has been displayed.",
             UserWarning,
+            stacklevel=3,
         )
         return ax
 
@@ -716,15 +772,18 @@ GON (((-122.84000 49.00000, -120.0000...
             figsize=figsize,
             markersize=markersize,
             aspect=aspect,
+            autolim=autolim,
             **style_kwds,
         )
 
     # To accept pd.Series and np.arrays as column
-    if isinstance(column, (np.ndarray, pd.Series)):
+    if isinstance(column, np.ndarray | pd.Series | pd.Index):
         if column.shape[0] != df.shape[0]:
             raise ValueError(
                 "The dataframe and given column have different number of rows."
             )
+        elif isinstance(column, pd.Index):
+            values = column.values
         else:
             values = column
 
@@ -734,7 +793,7 @@ GON (((-122.84000 49.00000, -120.0000...
     else:
         values = df[column]
 
-    if pd.api.types.is_categorical_dtype(values.dtype):
+    if isinstance(values.dtype, CategoricalDtype):
         if categories is not None:
             raise ValueError(
                 "Cannot specify 'categories' when column has categorical dtype"
@@ -751,17 +810,11 @@ GON (((-122.84000 49.00000, -120.0000...
     nan_idx = np.asarray(pd.isna(values), dtype="bool")
 
     if scheme is not None:
-        mc_err = (
-            "The 'mapclassify' package (>= 2.4.0) is "
-            "required to use the 'scheme' keyword."
-        )
+        mc_err = "The 'mapclassify' package is required to use the 'scheme' keyword."
         try:
             import mapclassify
 
         except ImportError:
-            raise ImportError(mc_err)
-
-        if Version(mapclassify.__version__) < Version("2.4.0"):
             raise ImportError(mc_err)
 
         if classification_kwds is None:
@@ -810,17 +863,12 @@ GON (((-122.84000 49.00000, -120.0000...
     if categorical:
         if cmap is None:
             cmap = "tab10"
-
-        cat = pd.Categorical(values, categories=categories)
+        cat = _check_invalid_categories(categories, values)
+        # if isinstance(values, pd.Categorical):
+        #     cat = values
+        # else:
+        #     cat = pd.Categorical(values, categories=categories)
         categories = list(cat.categories)
-
-        # values missing in the Categorical but not in original values
-        missing = list(np.unique(values[~nan_idx & cat.isna()]))
-        if missing:
-            raise ValueError(
-                "Column contains values not listed in categories. "
-                "Missing categories: {}.".format(missing)
-            )
 
         values = cat.codes[~nan_idx]
         vmin = 0 if vmin is None else vmin
@@ -836,10 +884,9 @@ GON (((-122.84000 49.00000, -120.0000...
     mx = values[~np.isnan(values)].max() if vmax is None else vmax
 
     # decompose GeometryCollections
-    geoms, multiindex = _sanitize_geoms(df.geometry, prefix="Geom")
+    expl_series, multiindex = _sanitize_geoms(df.geometry, prefix="Geom")
     values = np.take(values, multiindex, axis=0)
     nan_idx = np.take(nan_idx, multiindex, axis=0)
-    expl_series = geopandas.GeoSeries(geoms)
 
     geom_types = expl_series.geom_type
     poly_idx = np.asarray((geom_types == "Polygon") | (geom_types == "MultiPolygon"))
@@ -855,7 +902,14 @@ GON (((-122.84000 49.00000, -120.0000...
     subset = values[poly_idx & np.invert(nan_idx)]
     if not polys.empty:
         _plot_polygon_collection(
-            ax, polys, subset, vmin=mn, vmax=mx, cmap=cmap, **style_kwds
+            ax,
+            polys,
+            subset,
+            vmin=mn,
+            vmax=mx,
+            cmap=cmap,
+            autolim=autolim,
+            **style_kwds,
         )
 
     # plot all LineStrings and MultiLineString components in same collection
@@ -863,7 +917,14 @@ GON (((-122.84000 49.00000, -120.0000...
     subset = values[line_idx & np.invert(nan_idx)]
     if not lines.empty:
         _plot_linestring_collection(
-            ax, lines, subset, vmin=mn, vmax=mx, cmap=cmap, **style_kwds
+            ax,
+            lines,
+            subset,
+            vmin=mn,
+            vmax=mx,
+            cmap=cmap,
+            autolim=autolim,
+            **style_kwds,
         )
 
     # plot all Points in the same collection
@@ -893,7 +954,7 @@ GON (((-122.84000 49.00000, -120.0000...
         merged_kwds = style_kwds.copy()
         merged_kwds.update(missing_kwds)
 
-        plot_series(expl_series[nan_idx], ax=ax, **merged_kwds)
+        plot_series(expl_series[nan_idx], ax=ax, **merged_kwds, aspect=None)
 
     if legend and not color:
         if legend_kwds is None:
@@ -901,9 +962,9 @@ GON (((-122.84000 49.00000, -120.0000...
         if "fmt" in legend_kwds:
             legend_kwds.pop("fmt")
 
-        from matplotlib.lines import Line2D
-        from matplotlib.colors import Normalize
         from matplotlib import cm
+        from matplotlib.colors import Normalize
+        from matplotlib.lines import Line2D
 
         norm = style_kwds.get("norm", None)
         if not norm:
@@ -913,7 +974,7 @@ GON (((-122.84000 49.00000, -120.0000...
             if scheme is not None:
                 categories = labels
             patches = []
-            for value, cat in enumerate(categories):
+            for i in range(len(categories)):
                 patches.append(
                     Line2D(
                         [0],
@@ -922,7 +983,7 @@ GON (((-122.84000 49.00000, -120.0000...
                         marker="o",
                         alpha=style_kwds.get("alpha", 1),
                         markersize=10,
-                        markerfacecolor=n_cmap.to_rgba(value),
+                        markerfacecolor=n_cmap.to_rgba(i),
                         markeredgewidth=0,
                     )
                 )
@@ -947,7 +1008,9 @@ GON (((-122.84000 49.00000, -120.0000...
                 categories.append(merged_kwds.get("label", "NaN"))
             legend_kwds.setdefault("numpoints", 1)
             legend_kwds.setdefault("loc", "best")
-            ax.legend(patches, categories, **legend_kwds)
+            legend_kwds.setdefault("handles", patches)
+            legend_kwds.setdefault("labels", categories)
+            ax.legend(**legend_kwds)
         else:
             if cax is not None:
                 legend_kwds.setdefault("cax", cax)
@@ -957,8 +1020,37 @@ GON (((-122.84000 49.00000, -120.0000...
             n_cmap.set_array(np.array([]))
             ax.get_figure().colorbar(n_cmap, **legend_kwds)
 
-    plt.draw()
+    ax.figure.canvas.draw_idle()
     return ax
+
+
+def _check_invalid_categories(
+    categories: Collection[Any] | None, values
+) -> pd.Categorical:
+    if categories is None:
+        cat = pd.Categorical(values, categories=categories)
+    else:
+        # Pandas 4 compat https://github.com/pandas-dev/pandas/pull/62142
+        # Could potentially be replaced with a try/except on the above once the warning
+        # becomes an exception. This logic is derived from
+        # pandas/core/arrays/categorical.py::_get_codes_for_values
+        dtype = CategoricalDtype._from_values_or_dtype(values, categories)
+        categories = dtype.categories
+        codes = categories.get_indexer_for(values)
+        wrong = (codes == -1) & ~pd.isna(values)
+        if wrong.any():
+            missing = list(np.unique(values[wrong]))
+        else:
+            missing = []
+            codes_downcast = pd.core.dtypes.cast.coerce_indexer_dtype(codes, categories)
+            cat = pd.Categorical.from_codes(codes_downcast, categories)
+
+        if missing:
+            raise ValueError(
+                "Column contains values not listed in categories. "
+                f"Missing categories: {missing}."
+            )
+    return cat
 
 
 @doc(plot_dataframe)
@@ -978,4 +1070,4 @@ class GeoplotAccessor(PlotAccessor):
             raise ValueError(f"{kind} is not a valid plot kind")
 
     def geo(self, *args, **kwargs):
-        return self(kind="geo", *args, **kwargs)
+        return self(kind="geo", *args, **kwargs)  # noqa: B026
